Adjusted path and namespace to Entities.Constants
