--- managed/CounterStrikeSharp.API/Modules/Utils/ItemDefinition.cs
@@ -1,1724 +1,0 @@
-﻿/*
- *  This file is part of CounterStrikeSharp.
- *  CounterStrikeSharp is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  CounterStrikeSharp is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with CounterStrikeSharp.  If not, see <https://www.gnu.org/licenses/>. *
- */
-
-namespace CounterStrikeSharp.API.Modules.Utils;
-
-public enum ItemDefinition : ushort
-{
-    NONE,
-    DESERT_EAGLE,
-    DUAL_BERETTAS,
-    FIVE_SEVEN,
-    GLOCK_18,
-    AK_47 = 7,
-    AUG,
-    AWP,
-    FAMAS,
-    G3SG1,
-    GALIL_AR = 13,
-    M249,
-    M4A4 = 16,
-    MAC_10,
-    P90 = 19,
-    REPULSOR_DEVICE,
-    MP5_SD = 23,
-    UMP_45,
-    XM1014,
-    PP_BIZON,
-    MAG_7,
-    NEGEV,
-    SAWED_OFF,
-    TEC_9,
-    ZEUS_X27,
-    P2000,
-    MP7,
-    MP9,
-    NOVA,
-    P250,
-    RIOT_SHIELD,
-    SCAR_20,
-    SG_553,
-    SSG_08,
-    KNIFE_GG,
-    KNIFE_CT,
-    FLASHBANG,
-    HIGH_EXPLOSIVE_GRENADE,
-    SMOKE_GRENADE,
-    MOLOTOV,
-    DECOY_GRENADE,
-    INCENDIARY_GRENADE,
-    C4_EXPLOSIVE,
-    KEVLAR_VEST,
-    KEVLAR_AND_HELMET,
-    HEAVY_ASSAULT_SUIT,
-    NO_LOCALIZED_NAME0 = 54,
-    DEFUSE_KIT,
-    RESCUE_KIT,
-    HEALTHSHOT,
-    MUSIC_KIT,
-    KNIFE_T,
-    M4A1_S,
-    USP_S,
-    TRADE_UP_CONTRACT,
-    CZ75_AUTO,
-    R8_REVOLVER,
-    TACTICAL_AWARENESS_GRENADE = 68,
-    BARE_HANDS,
-    BREACH_CHARGE,
-    TABLET = 72,
-    MEELE = 74,
-    AXE,
-    HAMMER,
-    WRENCH = 78,
-    KNIFE_GHOST = 80,
-    FIRE_BOMB,
-    DIVERSION_DEVICE,
-    FRAG_GRENADE,
-    SNOWBALL,
-    BUMP_MINE,
-    BAYONET = 500,
-    CLASSIC_KNIFE = 503,
-    FLIP_KNIFE = 505,
-    GUT_KNIFE,
-    KARAMBIT,
-    M9_BAYONET,
-    HUNTSMAN_KNIFE,
-    FALCHION_KNIFE = 512,
-    BOWIE_KNIFE = 514,
-    BUTTERFLY_KNIFE,
-    SHADOW_DAGGERS,
-    PARACORD_KNIFE,
-    SURVIVAL_KNIFE,
-    URSUS_KNIFE,
-    NAVAJA_KNIFE,
-    NOMAD_KNIFE,
-    STILETTO_KNIFE,
-    TALON_KNIFE,
-    SKELETON_KNIFE = 525,
-    _5_YEAR_VETERAN_COIN = 874,
-    CHAMPION_AT_DREAMHACK_2013,
-    FINALIST_AT_DREAMHACK_2013,
-    SEMIFINALIST_AT_DREAMHACK_2013,
-    QUARTERFINALIST_AT_DREAMHACK_2013,
-    CHAMPION_AT_EMS_ONE_KATOWICE_2014,
-    FINALIST_AT_EMS_ONE_KATOWICE_2014,
-    SEMIFINALIST_AT_EMS_ONE_KATOWICE_2014,
-    QUARTERFINALIST_AT_EMS_ONE_KATOWICE_2014,
-    CHAMPION_AT_ESL_ONE_COLOGNE_2014,
-    FINALIST_AT_ESL_ONE_COLOGNE_2014,
-    SEMIFINALIST_AT_ESL_ONE_COLOGNE_2014,
-    QUARTERFINALIST_AT_ESL_ONE_COLOGNE_2014,
-    BRONZE_COLOGNE_2014_PICKEM_TROPHY,
-    SILVER_COLOGNE_2014_PICKEM_TROPHY,
-    GOLD_COLOGNE_2014_PICKEM_TROPHY,
-    CHAMPION_AT_DREAMHACK_WINTER_2014,
-    FINALIST_AT_DREAMHACK_WINTER_2014,
-    SEMIFINALIST_AT_DREAMHACK_WINTER_2014,
-    QUARTERFINALIST_AT_DREAMHACK_WINTER_2014,
-    BRONZE_DREAMHACK_2014_PICKEM_TROPHY,
-    SILVER_DREAMHACK_2014_PICKEM_TROPHY,
-    GOLD_DREAMHACK_2014_PICKEM_TROPHY,
-    CHAMPION_AT_ESL_ONE_KATOWICE_2015,
-    FINALIST_AT_ESL_ONE_KATOWICE_2015,
-    SEMIFINALIST_AT_ESL_ONE_KATOWICE_2015,
-    QUARTERFINALIST_AT_ESL_ONE_KATOWICE_2015,
-    BRONZE_KATOWICE_2015_PICKEM_TROPHY,
-    SILVER_KATOWICE_2015_PICKEM_TROPHY,
-    GOLD_KATOWICE_2015_PICKEM_TROPHY,
-    CHAMPION_AT_ESL_ONE_COLOGNE_2015,
-    FINALIST_AT_ESL_ONE_COLOGNE_2015,
-    SEMIFINALIST_AT_ESL_ONE_COLOGNE_2015,
-    QUARTERFINALIST_AT_ESL_ONE_COLOGNE_2015,
-    BRONZE_COLOGNE_2015_PICKEM_TROPHY,
-    SILVER_COLOGNE_2015_PICKEM_TROPHY,
-    GOLD_COLOGNE_2015_PICKEM_TROPHY,
-    BRONZE_CLUJ_NAPOCA_2015_PICKEM_TROPHY,
-    SILVER_CLUJ_NAPOCA_2015_PICKEM_TROPHY,
-    GOLD_CLUJ_NAPOCA_2015_PICKEM_TROPHY,
-    BRONZE_CLUJ_NAPOCA_2015_FANTASY_TROPHY,
-    SILVER_CLUJ_NAPOCA_2015_FANTASY_TROPHY,
-    GOLD_CLUJ_NAPOCA_2015_FANTASY_TROPHY,
-    CHAMPION_AT_DREAMHACK_CLUJ_NAPOCA_2015,
-    FINALIST_AT_DREAMHACK_CLUJ_NAPOCA_2015,
-    SEMIFINALIST_AT_DREAMHACK_CLUJ_NAPOCA_2015,
-    QUARTERFINALIST_AT_DREAMHACK_CLUJ_NAPOCA_2015,
-    BRONZE_COLUMBUS_2016_PICKEM_TROPHY,
-    SILVER_COLUMBUS_2016_PICKEM_TROPHY,
-    GOLD_COLUMBUS_2016_PICKEM_TROPHY,
-    BRONZE_COLUMBUS_2016_FANTASY_TROPHY,
-    SILVER_COLUMBUS_2016_FANTASY_TROPHY,
-    GOLD_COLUMBUS_2016_FANTASY_TROPHY,
-    CHAMPION_AT_MLG_COLUMBUS_2016,
-    FINALIST_AT_MLG_COLUMBUS_2016,
-    SEMIFINALIST_AT_MLG_COLUMBUS_2016,
-    QUARTERFINALIST_AT_MLG_COLUMBUS_2016,
-    CHAMPION_AT_ESL_ONE_COLOGNE_2016,
-    FINALIST_AT_ESL_ONE_COLOGNE_2016,
-    SEMIFINALIST_AT_ESL_ONE_COLOGNE_2016,
-    QUARTERFINALIST_AT_ESL_ONE_COLOGNE_2016,
-    BRONZE_COLOGNE_2016_PICKEM_TROPHY,
-    SILVER_COLOGNE_2016_PICKEM_TROPHY,
-    GOLD_COLOGNE_2016_PICKEM_TROPHY,
-    BRONZE_COLOGNE_2016_FANTASY_TROPHY,
-    SILVER_COLOGNE_2016_FANTASY_TROPHY,
-    GOLD_COLOGNE_2016_FANTASY_TROPHY,
-    CHAMPION_AT_ELEAGUE_ATLANTA_2017,
-    FINALIST_AT_ELEAGUE_ATLANTA_2017,
-    SEMIFINALIST_AT_ELEAGUE_ATLANTA_2017,
-    QUARTERFINALIST_AT_ELEAGUE_ATLANTA_2017,
-    BRONZE_ATLANTA_2017_PICKEM_TROPHY,
-    SILVER_ATLANTA_2017_PICKEM_TROPHY,
-    GOLD_ATLANTA_2017_PICKEM_TROPHY,
-    CHAMPION_AT_PGL_KRAKOW_2017,
-    FINALIST_AT_PGL_KRAKOW_2017,
-    SEMIFINALIST_AT_PGL_KRAKOW_2017,
-    QUARTERFINALIST_AT_PGL_KRAKOW_2017,
-    BRONZE_KRAKOW_2017_PICKEM_TROPHY,
-    SILVER_KRAKOW_2017_PICKEM_TROPHY,
-    GOLD_KRAKOW_2017_PICKEM_TROPHY,
-    CHAMPION_AT_ELEAGUE_BOSTON_2018,
-    FINALIST_AT_ELEAGUE_BOSTON_2018,
-    SEMIFINALIST_AT_ELEAGUE_BOSTON_2018,
-    QUARTERFINALIST_AT_ELEAGUE_BOSTON_2018,
-    BRONZE_BOSTON_2018_PICKEM_TROPHY,
-    SILVER_BOSTON_2018_PICKEM_TROPHY,
-    GOLD_BOSTON_2018_PICKEM_TROPHY,
-    CHAMPION_AT_FACEIT_LONDON_2018,
-    FINALIST_AT_FACEIT_LONDON_2018,
-    SEMIFINALIST_AT_FACEIT_LONDON_2018,
-    QUARTERFINALIST_AT_FACEIT_LONDON_2018,
-    BRONZE_LONDON_2018_PICKEM_TROPHY,
-    SILVER_LONDON_2018_PICKEM_TROPHY,
-    GOLD_LONDON_2018_PICKEM_TROPHY,
-    _10_YEAR_VETERAN_COIN,
-    LOYALTY_BADGE,
-    CHAMPION_AT_IEM_KATOWICE_2019,
-    FINALIST_AT_IEM_KATOWICE_2019,
-    SEMIFINALIST_AT_IEM_KATOWICE_2019,
-    QUARTERFINALIST_AT_IEM_KATOWICE_2019,
-    CHAMPION_AT_STARLADDER_BERLIN_2019,
-    FINALIST_AT_STARLADDER_BERLIN_2019,
-    SEMIFINALIST_AT_STARLADDER_BERLIN_2019,
-    QUARTERFINALIST_AT_STARLADDER_BERLIN_2019,
-    CHAMPION_AT_PGL_STOCKHOLM_2021,
-    FINALIST_AT_PGL_STOCKHOLM_2021,
-    SEMIFINALIST_AT_PGL_STOCKHOLM_2021,
-    QUARTERFINALIST_AT_PGL_STOCKHOLM_2021,
-    CHAMPION_AT_PGL_ANTWERP_2022,
-    FINALIST_AT_PGL_ANTWERP_2022,
-    SEMIFINALIST_AT_PGL_ANTWERP_2022,
-    QUARTERFINALIST_AT_PGL_ANTWERP_2022,
-    _10_YEAR_BIRTHDAY_COIN,
-    CHAMPION_AT_IEM_RIO_2022,
-    FINALIST_AT_IEM_RIO_2022,
-    SEMIFINALIST_AT_IEM_RIO_2022,
-    QUARTERFINALIST_AT_IEM_RIO_2022,
-    CHAMPION_AT_BLASTTV_PARIS_2023,
-    FINALIST_AT_BLASTTV_PARIS_2023,
-    SEMIFINALIST_AT_BLASTTV_PARIS_2023,
-    QUARTERFINALIST_AT_BLASTTV_PARIS_2023,
-    OPERATION_PAYBACK_PASS = 1000,
-    OPERATION_PAYBACK_CHALLENGE_COIN,
-    SILVER_OPERATION_PAYBACK_COIN,
-    GOLD_OPERATION_PAYBACK_COIN,
-    MUSEUM_MAP_COIN,
-    DOWNTOWN_MAP_COIN,
-    THUNDER_MAP_COIN,
-    FAVELA_MAP_COIN,
-    MOTEL_MAP_COIN,
-    SEASIDE_MAP_COIN,
-    LIBRARY_MAP_COIN,
-    OPERATION_BRAVO_PASS = 1012,
-    OPERATION_BRAVO_CHALLENGE_COIN,
-    SILVER_OPERATION_BRAVO_COIN,
-    GOLD_OPERATION_BRAVO_COIN,
-    AGENCY_MAP_COIN,
-    ALI_MAP_COIN,
-    CACHE_MAP_COIN,
-    CHINATOWN_MAP_COIN,
-    GWALIOR_MAP_COIN,
-    RUINS_MAP_COIN,
-    SIEGE_MAP_COIN,
-    OPERATION_PHOENIX_PASS,
-    OPERATION_PHOENIX_CHALLENGE_COIN,
-    SILVER_OPERATION_PHOENIX_COIN,
-    GOLD_OPERATION_PHOENIX_COIN,
-    OPERATION_BREAKOUT_ALL_ACCESS_PASS,
-    OPERATION_BREAKOUT_CHALLENGE_COIN,
-    SILVER_OPERATION_BREAKOUT_COIN,
-    GOLD_OPERATION_BREAKOUT_COIN,
-    CASTLE_MAP_COIN,
-    BLACK_GOLD_MAP_COIN,
-    RUSH_MAP_COIN,
-    MIST_MAP_COIN,
-    INSERTION_MAP_COIN,
-    OVERGROWN_MAP_COIN,
-    MARQUIS_MAP_COIN,
-    WORKOUT_MAP_COIN,
-    BACKALLEY_MAP_COIN,
-    SEASON_MAP_COIN,
-    BAZAAR_MAP_COIN,
-    FACADE_MAP_COIN,
-    LOG_MAP_COIN,
-    RAILS_MAP_COIN,
-    RESORT_MAP_COIN,
-    ZOO_MAP_COIN,
-    SANTORINI_MAP_COIN,
-    COAST_MAP_COIN,
-    MIKLA_MAP_COIN,
-    ROYAL_MAP_COIN,
-    EMPIRE_MAP_COIN,
-    TULIP_MAP_COIN,
-    CRUISE_MAP_COIN,
-    SUBZERO_MAP_COIN,
-    BIOME_MAP_COIN,
-    ABBEY_MAP_COIN,
-    RUBY_MAP_COIN,
-    BREACH_MAP_COIN,
-    STUDIO_MAP_COIN,
-    JUNGLE_MAP_COIN,
-    ANUBIS_MAP_COIN,
-    CHLORINE_MAP_COIN,
-    MUTINY_MAP_COIN,
-    SWAMP_MAP_COIN,
-    FROSTBITE_MAP_COIN,
-    ENGAGE_MAP_COIN,
-    APOLLO_MAP_COIN,
-    GUARD_MAP_COIN,
-    ELYSION_MAP_COIN,
-    GRIND_MAP_COIN,
-    MOCHA_MAP_COIN,
-    CALAVERA_MAP_COIN,
-    PITSTOP_MAP_COIN,
-    BASALT_MAP_COIN,
-    INSERTION_II_MAP_COIN,
-    RAVINE_MAP_COIN,
-    EXTRACTION_MAP_COIN,
-    COUNTY_MAP_COIN,
-    IRIS_MAP_COIN,
-    CLIMB_MAP_COIN,
-    CRETE_MAP_COIN,
-    HIVE_MAP_COIN,
-    VINEYARD_MAP_COIN,
-    EMBER_MAP_COIN,
-    TUSCAN_MAP_COIN,
-    CASCADE_MAP_COIN,
-    BLAGAI_MAP_COIN,
-    BOYARD_MAP_COIN,
-    CHALICE_MAP_COIN,
-    NAME_TAG = 1200,
-    STORAGE_UNIT,
-    CSGO_CASE_KEY = 1203,
-    ESPORTS_KEY,
-    STICKER = 1209,
-    GIFT_PACKAGE,
-    PALLET_OF_PRESENTS,
-    CSGO_CAPSULE_KEY,
-    WINTER_OFFENSIVE_CASE_KEY = 1214,
-    AUDIENCE_PARTICIPATION_PARCEL,
-    OPERATION_PHOENIX_CASE_KEY = 1303,
-    COMMUNITY_STICKER_CAPSULE_1_KEY,
-    HUNTSMAN_CASE_KEY,
-    MISSION,
-    HUNTSMAN_CASE_KEY1,
-    COMMUNITY_STICKER_CAPSULE_1_KEY1,
-    OPERATION_BREAKOUT_CASE_KEY,
-    OPERATION_BREAKOUT_CASE_KEY1,
-    OPERATION_BREAKOUT_CASE_KEY2,
-    OPERATION_BREAKOUT_CASE_KEY3 = 1313,
-    MUSIC_KIT1,
-    OPERATION_VANGUARD_ACCESS_PASS,
-    OPERATION_VANGUARD_CHALLENGE_COIN,
-    SILVER_OPERATION_VANGUARD_COIN,
-    GOLD_OPERATION_VANGUARD_COIN,
-    CAMPAIGN_MAGHREB = 1320,
-    CAMPAIGN_EURASIA,
-    OPERATION_VANGUARD_CASE_KEY,
-    CHROMA_CASE_KEY,
-    STATTRAK_SWAP_TOOL,
-    CHROMA_2_CASE_KEY,
-    OPERATION_BLOODHOUND_ACCESS_PASS,
-    OPERATION_BLOODHOUND_CHALLENGE_COIN,
-    SILVER_OPERATION_BLOODHOUND_COIN,
-    GOLD_OPERATION_BLOODHOUND_COIN,
-    FALCHION_CASE_KEY,
-    _2015_SERVICE_MEDAL,
-    _2015_SERVICE_MEDAL1,
-    SHADOW_CASE_KEY,
-    REVOLVER_CASE_KEY,
-    OPERATION_WILDFIRE_ACCESS_PASS,
-    OPERATION_WILDFIRE_CHALLENGE_COIN,
-    SILVER_OPERATION_WILDFIRE_COIN,
-    GOLD_OPERATION_WILDFIRE_COIN,
-    _2016_SERVICE_MEDAL,
-    _2016_SERVICE_MEDAL1,
-    _2016_SERVICE_MEDAL2,
-    _2016_SERVICE_MEDAL3,
-    _2016_SERVICE_MEDAL4,
-    _2016_SERVICE_MEDAL5,
-    CHROMA_3_CASE_KEY = 1347,
-    SEALED_GRAFFITI,
-    GRAFFITI,
-    GAMMA_CASE_KEY,
-    GAMMA_2_CASE_KEY,
-    OPERATION_HYDRA_ACCESS_PASS,
-    COUNTER_STRIKE_GAME_LICENSE,
-    GLOVE_CASE_KEY = 1356,
-    _2017_SERVICE_MEDAL,
-    _2017_SERVICE_MEDAL1,
-    _2017_SERVICE_MEDAL2,
-    _2017_SERVICE_MEDAL3,
-    _2017_SERVICE_MEDAL4,
-    _2017_SERVICE_MEDAL5,
-    _2017_SERVICE_MEDAL6,
-    SPECTRUM_CASE_KEY,
-    OPERATION_HYDRA_CASE_KEY,
-    SPECTRUM_2_CASE_KEY,
-    _2018_SERVICE_MEDAL,
-    _2018_SERVICE_MEDAL1,
-    _2018_SERVICE_MEDAL2,
-    _2018_SERVICE_MEDAL3,
-    _2018_SERVICE_MEDAL4,
-    _2018_SERVICE_MEDAL5,
-    CLUTCH_CASE_KEY,
-    HORIZON_CASE_KEY,
-    DANGER_ZONE_CASE_KEY,
-    _2019_SERVICE_MEDAL,
-    _2019_SERVICE_MEDAL1,
-    _2019_SERVICE_MEDAL2,
-    _2019_SERVICE_MEDAL3,
-    _2019_SERVICE_MEDAL4,
-    _2019_SERVICE_MEDAL5,
-    PRISMA_CASE_KEY = 1383,
-    SHATTERED_WEB_CASE_KEY,
-    CS20_CASE_KEY,
-    PRISMA_2_CASE_KEY,
-    FRACTURE_CASE_KEY,
-    OPERATION_BROKEN_FANG_CASE_KEY,
-    SNAKEBITE_CASE_KEY,
-    OPERATION_RIPTIDE_CASE_KEY,
-    DREAMS_AND_NIGHTMARES_CASE_KEY,
-    RECOIL_CASE_KEY,
-    REVOLUTION_CASE_KEY,
-    CSGO_CASE = 4001,
-    ESPORTS_2013_CASE,
-    OPERATION_BRAVO_CASE,
-    CSGO_CASE_2,
-    ESPORTS_2013_WINTER_CASE,
-    DREAMHACK_2013_SOUVENIR_PACKAGE,
-    STICKER_CAPSULE,
-    WINTER_OFFENSIVE_CASE = 4009,
-    CSGO_CASE_3,
-    OPERATION_PHOENIX_CASE,
-    STICKER_CAPSULE_2,
-    EMS_ONE_2014_SOUVENIR_PACKAGE,
-    EMS_KATOWICE_2014_CHALLENGERS,
-    EMS_KATOWICE_2014_LEGENDS,
-    COMMUNITY_STICKER_CAPSULE_1,
-    HUNTSMAN_CASE,
-    OPERATION_BREAKOUT_CASE,
-    ESPORTS_2014_SUMMER_CASE,
-    ESL_ONE_COLOGNE_2014_LEGENDS,
-    ESL_ONE_COLOGNE_2014_CHALLENGERS,
-    ESL_ONE_COLOGNE_2014_DUST_II_SOUVENIR_PACKAGE,
-    ESL_ONE_COLOGNE_2014_INFERNO_SOUVENIR_PACKAGE,
-    ESL_ONE_COLOGNE_2014_MIRAGE_SOUVENIR_PACKAGE,
-    ESL_ONE_COLOGNE_2014_NUKE_SOUVENIR_PACKAGE,
-    ESL_ONE_COLOGNE_2014_CACHE_SOUVENIR_PACKAGE,
-    ESL_ONE_COLOGNE_2014_COBBLESTONE_SOUVENIR_PACKAGE,
-    ESL_ONE_COLOGNE_2014_OVERPASS_SOUVENIR_PACKAGE,
-    OPERATION_VANGUARD_CASE,
-    DREAMHACK_2014_LEGENDS_HOLO_FOIL,
-    DREAMHACK_2014_DUST_II_SOUVENIR_PACKAGE,
-    DREAMHACK_2014_INFERNO_SOUVENIR_PACKAGE,
-    DREAMHACK_2014_MIRAGE_SOUVENIR_PACKAGE,
-    DREAMHACK_2014_NUKE_SOUVENIR_PACKAGE,
-    DREAMHACK_2014_CACHE_SOUVENIR_PACKAGE,
-    DREAMHACK_2014_COBBLESTONE_SOUVENIR_PACKAGE,
-    DREAMHACK_2014_OVERPASS_SOUVENIR_PACKAGE,
-    FNATIC_DREAMHACK_2014,
-    CLOUD9_DREAMHACK_2014,
-    NINJAS_IN_PYJAMAS_DREAMHACK_2014 = 4041,
-    VIRTUSPRO_DREAMHACK_2014,
-    NATUS_VINCERE_DREAMHACK_2014,
-    TEAM_DIGNITAS_DREAMHACK_2014 = 4045,
-    BRAVADO_GAMING_DREAMHACK_2014,
-    ESC_GAMING_DREAMHACK_2014,
-    HELLRAISERS_DREAMHACK_2014,
-    MYXMG_DREAMHACK_2014,
-    IBUYPOWER_DREAMHACK_2014,
-    TEAM_LDLCCOM_DREAMHACK_2014,
-    PENTA_SPORTS_DREAMHACK_2014,
-    PLANETKEY_DYNAMICS_DREAMHACK_2014,
-    DREAMHACK_WINTER_2014,
-    _3DMAX_DREAMHACK_2014,
-    COPENHAGEN_WOLVES_DREAMHACK_2014,
-    DAT_TEAM_DREAMHACK_2014,
-    LONDON_CONSPIRACY_DREAMHACK_2014,
-    MOUSESPORTS_DREAMHACK_2014,
-    FLIPSID3_TACTICS_DREAMHACK_2014,
-    CHROMA_CASE,
-    _3DMAX_KATOWICE_2015,
-    CLOUD9_G2A_KATOWICE_2015,
-    COUNTER_LOGIC_GAMING_KATOWICE_2015,
-    FLIPSID3_TACTICS_KATOWICE_2015,
-    FNATIC_KATOWICE_2015,
-    HELLRAISERS_KATOWICE_2015,
-    KEYD_STARS_KATOWICE_2015,
-    LGB_ESPORTS_KATOWICE_2015,
-    NATUS_VINCERE_KATOWICE_2015,
-    NINJAS_IN_PYJAMAS__KATOWICE_2015,
-    PENTA_SPORTS__KATOWICE_2015,
-    TEAM_ENVYUS_KATOWICE_2015,
-    TSM_KINGUIN_KATOWICE_2015,
-    TITAN_KATOWICE_2015,
-    VIRTUSPRO_KATOWICE_2015,
-    VOX_EMINOR__KATOWICE_2015,
-    ESL_ONE_KATOWICE_2015,
-    ESL_ONE_KATOWICE_2015_DUST_II_SOUVENIR_PACKAGE,
-    ESL_ONE_KATOWICE_2015_INFERNO_SOUVENIR_PACKAGE,
-    ESL_ONE_KATOWICE_2015_MIRAGE_SOUVENIR_PACKAGE,
-    ESL_ONE_KATOWICE_2015_NUKE_SOUVENIR_PACKAGE,
-    ESL_ONE_KATOWICE_2015_CACHE_SOUVENIR_PACKAGE,
-    ESL_ONE_KATOWICE_2015_COBBLESTONE_SOUVENIR_PACKAGE,
-    ESL_ONE_KATOWICE_2015_OVERPASS_SOUVENIR_PACKAGE,
-    ESL_ONE_KATOWICE_2015_LEGENDS_HOLO_FOIL,
-    ESL_ONE_KATOWICE_2015_CHALLENGERS_HOLO_FOIL,
-    STATTRAK_SWAP_TOOL_TWO_PACK,
-    CHROMA_2_CASE,
-    ENFU_STICKER_CAPSULE,
-    FALCHION_CASE,
-    FNATIC_COLOGNE_2015,
-    VIRTUSPRO_COLOGNE_2015,
-    MOUSESPORTS_COLOGNE_2015,
-    NATUS_VINCERE_COLOGNE_2015,
-    RENEGADES_COLOGNE_2015,
-    TEAM_KINGUIN_COLOGNE_2015,
-    TEAM_EBETTLE_COLOGNE_2015,
-    CLOUD9_G2A_COLOGNE_2015,
-    NINJAS_IN_PYJAMAS_COLOGNE_2015,
-    TEAM_ENVYUS_COLOGNE_2015,
-    LUMINOSITY_GAMING_COLOGNE_2015,
-    TEAM_SOLOMID_COLOGNE_2015,
-    TEAM_IMMUNITY_COLOGNE_2015,
-    FLIPSID3_TACTICS_COLOGNE_2015,
-    TITAN_COLOGNE_2015,
-    COUNTER_LOGIC_GAMING_COLOGNE_2015,
-    ESL_COLOGNE_2015,
-    ESL_ONE_COLOGNE_2015_LEGENDS_FOIL,
-    ESL_ONE_COLOGNE_2015_CHALLENGERS_FOIL,
-    AUTOGRAPH_CAPSULE_GROUP_A_FOIL_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_GROUP_B_FOIL_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_GROUP_C_FOIL_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_GROUP_D_FOIL_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_FNATIC_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_LUMINOSITY_GAMING_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_NATUS_VINCERE_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_NINJAS_IN_PYJAMAS_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_TEAM_ENVYUS_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_TITAN_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_TEAM_SOLOMID_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_VIRTUSPRO_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_MOUSESPORTS_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_RENEGADES_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_TEAM_IMMUNITY_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_TEAM_EBETTLE_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_TEAM_KINGUIN_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_FLIPSID3_TACTICS_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_COUNTER_LOGIC_GAMING_COLOGNE_2015,
-    AUTOGRAPH_CAPSULE_CLOUD9_G2A_COLOGNE_2015,
-    ESL_ONE_COLOGNE_2015_DUST_II_SOUVENIR_PACKAGE,
-    ESL_ONE_COLOGNE_2015_MIRAGE_SOUVENIR_PACKAGE,
-    ESL_ONE_COLOGNE_2015_INFERNO_SOUVENIR_PACKAGE,
-    ESL_ONE_COLOGNE_2015_COBBLESTONE_SOUVENIR_PACKAGE,
-    ESL_ONE_COLOGNE_2015_OVERPASS_SOUVENIR_PACKAGE,
-    ESL_ONE_COLOGNE_2015_CACHE_SOUVENIR_PACKAGE,
-    ESL_ONE_COLOGNE_2015_TRAIN_SOUVENIR_PACKAGE,
-    SHADOW_CASE,
-    NINJAS_IN_PYJAMAS_CLUJ_NAPOCA_2015,
-    TEAM_DIGNITAS_CLUJ_NAPOCA_2015,
-    COUNTER_LOGIC_GAMING_CLUJ_NAPOCA_2015,
-    VEXED_GAMING_CLUJ_NAPOCA_2015,
-    FLIPSID3_TACTICS_CLUJ_NAPOCA_2015,
-    TEAM_LIQUID_CLUJ_NAPOCA_2015,
-    MOUSESPORTS_CLUJ_NAPOCA_2015,
-    NATUS_VINCERE_CLUJ_NAPOCA_2015,
-    VIRTUSPRO_CLUJ_NAPOCA_2015,
-    CLOUD9_CLUJ_NAPOCA_2015,
-    G2_ESPORTS_CLUJ_NAPOCA_2015,
-    TITAN_CLUJ_NAPOCA_2015,
-    TEAM_SOLOMID_CLUJ_NAPOCA_2015,
-    TEAM_ENVYUS_CLUJ_NAPOCA_2015,
-    FNATIC_CLUJ_NAPOCA_2015,
-    LUMINOSITY_GAMING_CLUJ_NAPOCA_2015,
-    DREAMHACK_CLUJ_NAPOCA_2015,
-    DREAMHACK_CLUJ_NAPOCA_2015_LEGENDS_FOIL,
-    DREAMHACK_CLUJ_NAPOCA_2015_CHALLENGERS_FOIL,
-    AUTOGRAPH_CAPSULE_CHALLENGERS_FOIL_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_LEGENDS_FOIL_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_NINJAS_IN_PYJAMAS_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_TEAM_DIGNITAS_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_COUNTER_LOGIC_GAMING_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_VEXED_GAMING_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_FLIPSID3_TACTICS_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_TEAM_LIQUID_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_MOUSESPORTS_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_NATUS_VINCERE_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_VIRTUSPRO_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_CLOUD9_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_G2_ESPORTS_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_TITAN_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_TEAM_SOLOMID_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_TEAM_ENVYUS_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_FNATIC_CLUJ_NAPOCA_2015,
-    AUTOGRAPH_CAPSULE_LUMINOSITY_GAMING_CLUJ_NAPOCA_2015,
-    DREAMHACK_CLUJ_NAPOCA_2015_DUST_II_SOUVENIR_PACKAGE,
-    DREAMHACK_CLUJ_NAPOCA_2015_MIRAGE_SOUVENIR_PACKAGE,
-    DREAMHACK_CLUJ_NAPOCA_2015_INFERNO_SOUVENIR_PACKAGE,
-    DREAMHACK_CLUJ_NAPOCA_2015_COBBLESTONE_SOUVENIR_PACKAGE,
-    DREAMHACK_CLUJ_NAPOCA_2015_OVERPASS_SOUVENIR_PACKAGE,
-    DREAMHACK_CLUJ_NAPOCA_2015_CACHE_SOUVENIR_PACKAGE,
-    DREAMHACK_CLUJ_NAPOCA_2015_TRAIN_SOUVENIR_PACKAGE,
-    PINUPS_CAPSULE,
-    SLID3_CAPSULE,
-    TEAM_ROLES_CAPSULE,
-    REVOLVER_CASE,
-    OPERATION_WILDFIRE_CASE,
-    NINJAS_IN_PYJAMAS_MLG_COLUMBUS_2016,
-    SPLYCE_MLG_COLUMBUS_2016,
-    COUNTER_LOGIC_GAMING_MLG_COLUMBUS_2016,
-    GAMBIT_GAMING_MLG_COLUMBUS_2016,
-    FLIPSID3_TACTICS_MLG_COLUMBUS_2016,
-    TEAM_LIQUID_MLG_COLUMBUS_2016,
-    MOUSESPORTS_MLG_COLUMBUS_2016,
-    NATUS_VINCERE_MLG_COLUMBUS_2016,
-    VIRTUSPRO_MLG_COLUMBUS_2016,
-    CLOUD9_MLG_COLUMBUS_2016,
-    G2_ESPORTS_MLG_COLUMBUS_2016,
-    FAZE_CLAN_MLG_COLUMBUS_2016,
-    ASTRALIS_MLG_COLUMBUS_2016,
-    TEAM_ENVYUS_MLG_COLUMBUS_2016,
-    FNATIC_MLG_COLUMBUS_2016,
-    LUMINOSITY_GAMING_MLG_COLUMBUS_2016,
-    MLG_MLG_COLUMBUS_2016,
-    MLG_COLUMBUS_2016_LEGENDS_HOLO_FOIL,
-    MLG_COLUMBUS_2016_CHALLENGERS_HOLO_FOIL,
-    AUTOGRAPH_CAPSULE_CHALLENGERS_FOIL_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_LEGENDS_FOIL_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_NINJAS_IN_PYJAMAS_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_SPLYCE_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_COUNTER_LOGIC_GAMING_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_GAMBIT_GAMING_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_FLIPSID3_TACTICS_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_TEAM_LIQUID_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_MOUSESPORTS_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_NATUS_VINCERE_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_VIRTUSPRO_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_CLOUD9_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_G2_ESPORTS_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_FAZE_CLAN_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_ASTRALIS_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_TEAM_ENVYUS_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_FNATIC_MLG_COLUMBUS_2016,
-    AUTOGRAPH_CAPSULE_LUMINOSITY_GAMING_MLG_COLUMBUS_2016,
-    MLG_COLUMBUS_2016_DUST_II_SOUVENIR_PACKAGE,
-    MLG_COLUMBUS_2016_MIRAGE_SOUVENIR_PACKAGE,
-    MLG_COLUMBUS_2016_INFERNO_SOUVENIR_PACKAGE,
-    MLG_COLUMBUS_2016_COBBLESTONE_SOUVENIR_PACKAGE,
-    MLG_COLUMBUS_2016_OVERPASS_SOUVENIR_PACKAGE,
-    MLG_COLUMBUS_2016_CACHE_SOUVENIR_PACKAGE,
-    MLG_COLUMBUS_2016_TRAIN_SOUVENIR_PACKAGE,
-    MLG_COLUMBUS_2016_NUKE_SOUVENIR_PACKAGE,
-    CHROMA_3_CASE,
-    COMMUNITY_GRAFFITI_BOX_1,
-    COLLECTIBLE_PINS_CAPSULE_SERIES_1,
-    GAMMA_CASE,
-    NINJAS_IN_PYJAMAS_COLOGNE_2016,
-    OPTIC_GAMING_COLOGNE_2016,
-    COUNTER_LOGIC_GAMING_COLOGNE_2016,
-    GAMBIT_GAMING_COLOGNE_2016,
-    FLIPSID3_TACTICS_COLOGNE_2016,
-    TEAM_LIQUID_COLOGNE_2016,
-    MOUSESPORTS_COLOGNE_2016,
-    NATUS_VINCERE_COLOGNE_2016,
-    VIRTUSPRO_COLOGNE_2016,
-    SK_GAMING_COLOGNE_2016,
-    G2_ESPORTS_COLOGNE_2016,
-    FAZE_CLAN_COLOGNE_2016,
-    ASTRALIS_COLOGNE_2016,
-    TEAM_ENVYUS_COLOGNE_2016,
-    FNATIC_COLOGNE_2016,
-    TEAM_DIGNITAS_COLOGNE_2016,
-    ESL_COLOGNE_2016,
-    COLOGNE_2016_LEGENDS_HOLO_FOIL,
-    COLOGNE_2016_CHALLENGERS_HOLO_FOIL,
-    AUTOGRAPH_CAPSULE_CHALLENGERS_FOIL_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_LEGENDS_FOIL_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_NINJAS_IN_PYJAMAS_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_OPTIC_GAMING_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_COUNTER_LOGIC_GAMING_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_GAMBIT_GAMING_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_FLIPSID3_TACTICS_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_TEAM_LIQUID_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_MOUSESPORTS_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_NATUS_VINCERE_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_VIRTUSPRO_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_SK_GAMING_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_G2_ESPORTS_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_FAZE_CLAN_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_ASTRALIS_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_TEAM_ENVYUS_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_FNATIC_COLOGNE_2016,
-    AUTOGRAPH_CAPSULE_TEAM_DIGNITAS_COLOGNE_2016,
-    COLOGNE_2016_DUST_II_SOUVENIR_PACKAGE,
-    COLOGNE_2016_MIRAGE_SOUVENIR_PACKAGE,
-    COLOGNE_2016_COBBLESTONE_SOUVENIR_PACKAGE,
-    COLOGNE_2016_OVERPASS_SOUVENIR_PACKAGE,
-    COLOGNE_2016_CACHE_SOUVENIR_PACKAGE,
-    COLOGNE_2016_TRAIN_SOUVENIR_PACKAGE,
-    COLOGNE_2016_NUKE_SOUVENIR_PACKAGE,
-    GAMMA_2_CASE,
-    SUGARFACE_CAPSULE,
-    BESTIARY_CAPSULE,
-    COLLECTIBLE_PINS_CAPSULE_SERIES_2,
-    CSGO_GRAFFITI_BOX,
-    PERFECT_WORLD_GRAFFITI_BOX,
-    STATTRAK_RADICALS_BOX,
-    GLOVE_CASE,
-    STICKER_ASTRALIS_ATLANTA_2017,
-    STICKER_TEAM_ENVYUS_ATLANTA_2017,
-    STICKER_FAZE_CLAN_ATLANTA_2017,
-    STICKER_FLIPSID3_TACTICS_ATLANTA_2017,
-    STICKER_FNATIC_ATLANTA_2017,
-    STICKER_G2_ESPORTS_ATLANTA_2017,
-    STICKER_GAMBIT_GAMING_ATLANTA_2017,
-    STICKER_GODSENT_ATLANTA_2017,
-    STICKER_HELLRAISERS_ATLANTA_2017,
-    STICKER_MOUSESPORTS_ATLANTA_2017,
-    STICKER_NATUS_VINCERE_ATLANTA_2017,
-    STICKER_NORTH_ATLANTA_2017,
-    STICKER_OPTIC_GAMING_ATLANTA_2017,
-    STICKER_SK_GAMING_ATLANTA_2017,
-    STICKER_TEAM_LIQUID_ATLANTA_2017,
-    STICKER_VIRTUSPRO_ATLANTA_2017,
-    STICKER_ELEAGUE_ATLANTA_2017,
-    SEALED_GRAFFITI_ASTRALIS_ATLANTA_2017,
-    SEALED_GRAFFITI_TEAM_ENVYUS_ATLANTA_2017,
-    SEALED_GRAFFITI_FAZE_CLAN_ATLANTA_2017,
-    SEALED_GRAFFITI_FLIPSID3_TACTICS_ATLANTA_2017,
-    SEALED_GRAFFITI_FNATIC_ATLANTA_2017,
-    SEALED_GRAFFITI_G2_ESPORTS_ATLANTA_2017,
-    SEALED_GRAFFITI_GAMBIT_GAMING_ATLANTA_2017,
-    SEALED_GRAFFITI_GODSENT_ATLANTA_2017,
-    SEALED_GRAFFITI_HELLRAISERS_ATLANTA_2017,
-    SEALED_GRAFFITI_MOUSESPORTS_ATLANTA_2017,
-    SEALED_GRAFFITI_NATUS_VINCERE_ATLANTA_2017,
-    SEALED_GRAFFITI_NORTH_ATLANTA_2017,
-    SEALED_GRAFFITI_OPTIC_GAMING_ATLANTA_2017,
-    SEALED_GRAFFITI_SK_GAMING_ATLANTA_2017,
-    SEALED_GRAFFITI_TEAM_LIQUID_ATLANTA_2017,
-    SEALED_GRAFFITI_VIRTUSPRO_ATLANTA_2017,
-    SEALED_GRAFFITI_ELEAGUE_ATLANTA_2017,
-    ATLANTA_2017_LEGENDS_HOLO_FOIL,
-    ATLANTA_2017_CHALLENGERS_HOLO_FOIL,
-    AUTOGRAPH_CAPSULE_CHALLENGERS_FOIL_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_LEGENDS_FOIL_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_ASTRALIS_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_TEAM_ENVYUS_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_FAZE_CLAN_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_FLIPSID3_TACTICS_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_FNATIC_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_G2_ESPORTS_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_GAMBIT_GAMING_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_GODSENT_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_HELLRAISERS_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_MOUSESPORTS_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_NATUS_VINCERE_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_NORTH_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_OPTIC_GAMING_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_SK_GAMING_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_TEAM_LIQUID_ATLANTA_2017,
-    AUTOGRAPH_CAPSULE_VIRTUSPRO_ATLANTA_2017,
-    ATLANTA_2017_MEGA_BUNDLE,
-    ATLANTA_2017_DUST_II_SOUVENIR_PACKAGE,
-    ATLANTA_2017_MIRAGE_SOUVENIR_PACKAGE,
-    ATLANTA_2017_COBBLESTONE_SOUVENIR_PACKAGE,
-    ATLANTA_2017_OVERPASS_SOUVENIR_PACKAGE,
-    ATLANTA_2017_CACHE_SOUVENIR_PACKAGE,
-    ATLANTA_2017_TRAIN_SOUVENIR_PACKAGE,
-    ATLANTA_2017_NUKE_SOUVENIR_PACKAGE,
-    SPECTRUM_CASE,
-    OPERATION_HYDRA_CASE,
-    OPERATION_HYDRA_CHALLENGE_COIN,
-    SILVER_OPERATION_HYDRA_COIN,
-    GOLD_OPERATION_HYDRA_COIN,
-    DIAMOND_OPERATION_HYDRA_COIN,
-    STICKER_ASTRALIS_KRAKOW_2017,
-    STICKER_VIRTUSPRO_KRAKOW_2017,
-    STICKER_FNATIC_KRAKOW_2017,
-    STICKER_SK_GAMING_KRAKOW_2017,
-    STICKER_NATUS_VINCERE_KRAKOW_2017,
-    STICKER_GAMBIT_KRAKOW_2017,
-    STICKER_NORTH_KRAKOW_2017,
-    STICKER_FAZE_CLAN_KRAKOW_2017,
-    STICKER_MOUSESPORTS_KRAKOW_2017,
-    STICKER_G2_ESPORTS_KRAKOW_2017,
-    STICKER_BIG_KRAKOW_2017,
-    STICKER_CLOUD9_KRAKOW_2017,
-    STICKER_PENTA_SPORTS_KRAKOW_2017,
-    STICKER_FLIPSID3_TACTICS_KRAKOW_2017,
-    STICKER_IMMORTALS_KRAKOW_2017,
-    STICKER_VEGA_SQUADRON_KRAKOW_2017,
-    STICKER_PGL_KRAKOW_2017,
-    SEALED_GRAFFITI_ASTRALIS_KRAKOW_2017,
-    SEALED_GRAFFITI_VIRTUSPRO_KRAKOW_2017,
-    SEALED_GRAFFITI_FNATIC_KRAKOW_2017,
-    SEALED_GRAFFITI_SK_GAMING_KRAKOW_2017,
-    SEALED_GRAFFITI_NATUS_VINCERE_KRAKOW_2017,
-    SEALED_GRAFFITI_GAMBIT_KRAKOW_2017,
-    SEALED_GRAFFITI_NORTH_KRAKOW_2017,
-    SEALED_GRAFFITI_FAZE_CLAN_KRAKOW_2017,
-    SEALED_GRAFFITI_MOUSESPORTS_KRAKOW_2017,
-    SEALED_GRAFFITI_G2_ESPORTS_KRAKOW_2017,
-    SEALED_GRAFFITI_BIG_KRAKOW_2017,
-    SEALED_GRAFFITI_CLOUD9_KRAKOW_2017,
-    SEALED_GRAFFITI_PENTA_SPORTS_KRAKOW_2017,
-    SEALED_GRAFFITI_FLIPSID3_TACTICS_KRAKOW_2017,
-    SEALED_GRAFFITI_IMMORTALS_KRAKOW_2017,
-    SEALED_GRAFFITI_VEGA_SQUADRON_KRAKOW_2017,
-    SEALED_GRAFFITI_PGL_KRAKOW_2017,
-    KRAKOW_2017_LEGENDS_HOLO_FOIL,
-    KRAKOW_2017_CHALLENGERS_HOLO_FOIL,
-    KRAKOW_2017_CHALLENGERS_AUTOGRAPH_CAPSULE,
-    KRAKOW_2017_LEGENDS_AUTOGRAPH_CAPSULE,
-    KRAKOW_2017_MEGA_BUNDLE,
-    KRAKOW_2017_INFERNO_SOUVENIR_PACKAGE,
-    KRAKOW_2017_MIRAGE_SOUVENIR_PACKAGE,
-    KRAKOW_2017_COBBLESTONE_SOUVENIR_PACKAGE,
-    KRAKOW_2017_OVERPASS_SOUVENIR_PACKAGE,
-    KRAKOW_2017_CACHE_SOUVENIR_PACKAGE,
-    KRAKOW_2017_TRAIN_SOUVENIR_PACKAGE,
-    KRAKOW_2017_NUKE_SOUVENIR_PACKAGE,
-    SPECTRUM_2_CASE,
-    PERFECT_WORLD_STICKER_CAPSULE_1,
-    PERFECT_WORLD_STICKER_CAPSULE_2,
-    STICKER_GAMBIT_ESPORTS_BOSTON_2018,
-    STICKER_100_THIEVES_BOSTON_2018,
-    STICKER_ASTRALIS_BOSTON_2018,
-    STICKER_VIRTUSPRO_BOSTON_2018,
-    STICKER_FNATIC_BOSTON_2018,
-    STICKER_SK_GAMING_BOSTON_2018,
-    STICKER_BIG_BOSTON_2018,
-    STICKER_NORTH_BOSTON_2018,
-    STICKER_G2_ESPORTS_BOSTON_2018,
-    STICKER_CLOUD9_BOSTON_2018,
-    STICKER_FLIPSID3_TACTICS_BOSTON_2018,
-    STICKER_NATUS_VINCERE_BOSTON_2018,
-    STICKER_MOUSESPORTS_BOSTON_2018,
-    STICKER_SPROUT_ESPORTS_BOSTON_2018,
-    STICKER_FAZE_CLAN_BOSTON_2018,
-    STICKER_VEGA_SQUADRON_BOSTON_2018,
-    STICKER_SPACE_SOLDIERS_BOSTON_2018,
-    STICKER_TEAM_LIQUID_BOSTON_2018,
-    STICKER_AVANGAR_BOSTON_2018,
-    STICKER_RENEGADES_BOSTON_2018,
-    STICKER_TEAM_ENVYUS_BOSTON_2018,
-    STICKER_MISFITS_GAMING_BOSTON_2018,
-    STICKER_QUANTUM_BELLATOR_FIRE_BOSTON_2018,
-    STICKER_TYLOO_BOSTON_2018,
-    STICKER_ELEAGUE_BOSTON_2018,
-    SEALED_GRAFFITI_GAMBIT_ESPORTS_BOSTON_2018,
-    SEALED_GRAFFITI_100_THIEVES_BOSTON_2018,
-    SEALED_GRAFFITI_ASTRALIS_BOSTON_2018,
-    SEALED_GRAFFITI_VIRTUSPRO_BOSTON_2018,
-    SEALED_GRAFFITI_FNATIC_BOSTON_2018,
-    SEALED_GRAFFITI_SK_GAMING_BOSTON_2018,
-    SEALED_GRAFFITI_BIG_BOSTON_2018,
-    SEALED_GRAFFITI_NORTH_BOSTON_2018,
-    SEALED_GRAFFITI_G2_ESPORTS_BOSTON_2018,
-    SEALED_GRAFFITI_CLOUD9_BOSTON_2018,
-    SEALED_GRAFFITI_FLIPSID3_TACTICS_BOSTON_2018,
-    SEALED_GRAFFITI_NATUS_VINCERE_BOSTON_2018,
-    SEALED_GRAFFITI_MOUSESPORTS_BOSTON_2018,
-    SEALED_GRAFFITI_SPROUT_ESPORTS_BOSTON_2018,
-    SEALED_GRAFFITI_FAZE_CLAN_BOSTON_2018,
-    SEALED_GRAFFITI_VEGA_SQUADRON_BOSTON_2018,
-    SEALED_GRAFFITI_SPACE_SOLDIERS_BOSTON_2018,
-    SEALED_GRAFFITI_TEAM_LIQUID_BOSTON_2018,
-    SEALED_GRAFFITI_AVANGAR_BOSTON_2018,
-    SEALED_GRAFFITI_RENEGADES_BOSTON_2018,
-    SEALED_GRAFFITI_TEAM_ENVYUS_BOSTON_2018,
-    SEALED_GRAFFITI_MISFITS_GAMING_BOSTON_2018,
-    SEALED_GRAFFITI_QUANTUM_BELLATOR_FIRE_BOSTON_2018,
-    SEALED_GRAFFITI_TYLOO_BOSTON_2018,
-    SEALED_GRAFFITI_ELEAGUE_BOSTON_2018,
-    BOSTON_2018_LEGENDS_HOLO_FOIL,
-    BOSTON_2018_RETURNING_CHALLENGERS_HOLO_FOIL,
-    BOSTON_2018_MINOR_CHALLENGERS_HOLO_FOIL,
-    BOSTON_2018_LEGENDS_AUTOGRAPH_CAPSULE,
-    BOSTON_2018_RETURNING_CHALLENGERS_AUTOGRAPH_CAPSULE,
-    BOSTON_2018_MINOR_CHALLENGERS_AUTOGRAPH_CAPSULE,
-    BOSTON_2018_MEGA_BUNDLE,
-    BOSTON_2018_INFERNO_SOUVENIR_PACKAGE,
-    BOSTON_2018_MIRAGE_SOUVENIR_PACKAGE,
-    BOSTON_2018_COBBLESTONE_SOUVENIR_PACKAGE,
-    BOSTON_2018_OVERPASS_SOUVENIR_PACKAGE,
-    BOSTON_2018_CACHE_SOUVENIR_PACKAGE,
-    BOSTON_2018_TRAIN_SOUVENIR_PACKAGE,
-    BOSTON_2018_NUKE_SOUVENIR_PACKAGE,
-    COMMUNITY_CAPSULE_2018,
-    CLUTCH_CASE,
-    STICKER_FLASH_GAMING_BOSTON_2018,
-    SEALED_GRAFFITI_FLASH_GAMING_BOSTON_2018,
-    BOSTON_2018_MINOR_CHALLENGERS_WITH_FLASH_GAMING_HOLO_FOIL,
-    BOSTON_2018_MINOR_CHALLENGERS_WITH_FLASH_GAMING_AUTOGRAPH_CAPSULE,
-    BOSTON_2018_MEGA_BUNDLE1,
-    CHICKEN_CAPSULE,
-    BOSTON_2018_ATTENDING_LEGENDS_HOLO_FOIL,
-    BOSTON_2018_ATTENDING_LEGENDS_AUTOGRAPH_CAPSULE,
-    BOSTON_2018_MEGA_BUNDLE2,
-    COLLECTIBLE_PINS_CAPSULE_SERIES_3,
-    HORIZON_CASE,
-    STICKER_CLOUD9_LONDON_2018,
-    STICKER_FAZE_CLAN_LONDON_2018,
-    STICKER_NATUS_VINCERE_LONDON_2018,
-    STICKER_MIBR_LONDON_2018,
-    STICKER_MOUSESPORTS_LONDON_2018,
-    STICKER_WINSTRIKE_TEAM_LONDON_2018,
-    STICKER_G2_ESPORTS_LONDON_2018,
-    STICKER_FNATIC_LONDON_2018,
-    STICKER_GAMBIT_ESPORTS_LONDON_2018,
-    STICKER_VEGA_SQUADRON_LONDON_2018,
-    STICKER_SPACE_SOLDIERS_LONDON_2018,
-    STICKER_BIG_LONDON_2018,
-    STICKER_ASTRALIS_LONDON_2018,
-    STICKER_TEAM_LIQUID_LONDON_2018,
-    STICKER_NORTH_LONDON_2018,
-    STICKER_VIRTUSPRO_LONDON_2018,
-    STICKER_NINJAS_IN_PYJAMAS_LONDON_2018,
-    STICKER_COMPLEXITY_GAMING_LONDON_2018,
-    STICKER_HELLRAISERS_LONDON_2018,
-    STICKER_RENEGADES_LONDON_2018,
-    STICKER_OPTIC_GAMING_LONDON_2018,
-    STICKER_ROGUE_LONDON_2018,
-    STICKER_TEAM_SPIRIT_LONDON_2018,
-    STICKER_TYLOO_LONDON_2018,
-    STICKER_FACEIT_LONDON_2018,
-    SEALED_GRAFFITI_CLOUD9_LONDON_2018,
-    SEALED_GRAFFITI_FAZE_CLAN_LONDON_2018,
-    SEALED_GRAFFITI_NATUS_VINCERE_LONDON_2018,
-    SEALED_GRAFFITI_MIBR_LONDON_2018,
-    SEALED_GRAFFITI_MOUSESPORTS_LONDON_2018,
-    SEALED_GRAFFITI_WINSTRIKE_TEAM_LONDON_2018,
-    SEALED_GRAFFITI_G2_ESPORTS_LONDON_2018,
-    SEALED_GRAFFITI_FNATIC_LONDON_2018,
-    SEALED_GRAFFITI_GAMBIT_ESPORTS_LONDON_2018,
-    SEALED_GRAFFITI_VEGA_SQUADRON_LONDON_2018,
-    SEALED_GRAFFITI_SPACE_SOLDIERS_LONDON_2018,
-    SEALED_GRAFFITI_BIG_LONDON_2018,
-    SEALED_GRAFFITI_ASTRALIS_LONDON_2018,
-    SEALED_GRAFFITI_TEAM_LIQUID_LONDON_2018,
-    SEALED_GRAFFITI_NORTH_LONDON_2018,
-    SEALED_GRAFFITI_VIRTUSPRO_LONDON_2018,
-    SEALED_GRAFFITI_NINJAS_IN_PYJAMAS_LONDON_2018,
-    SEALED_GRAFFITI_COMPLEXITY_GAMING_LONDON_2018,
-    SEALED_GRAFFITI_HELLRAISERS_LONDON_2018,
-    SEALED_GRAFFITI_RENEGADES_LONDON_2018,
-    SEALED_GRAFFITI_OPTIC_GAMING_LONDON_2018,
-    SEALED_GRAFFITI_ROGUE_LONDON_2018,
-    SEALED_GRAFFITI_TEAM_SPIRIT_LONDON_2018,
-    SEALED_GRAFFITI_TYLOO_LONDON_2018,
-    SEALED_GRAFFITI_FACEIT_LONDON_2018,
-    LONDON_2018_LEGENDS_HOLO_FOIL,
-    LONDON_2018_RETURNING_CHALLENGERS_HOLO_FOIL,
-    LONDON_2018_MINOR_CHALLENGERS_HOLO_FOIL,
-    LONDON_2018_LEGENDS_AUTOGRAPH_CAPSULE,
-    LONDON_2018_RETURNING_CHALLENGERS_AUTOGRAPH_CAPSULE,
-    LONDON_2018_MINOR_CHALLENGERS_AUTOGRAPH_CAPSULE,
-    LONDON_2018_MEGA_BUNDLE,
-    LONDON_2018_INFERNO_SOUVENIR_PACKAGE,
-    LONDON_2018_MIRAGE_SOUVENIR_PACKAGE,
-    LONDON_2018_DUST_II_SOUVENIR_PACKAGE,
-    LONDON_2018_OVERPASS_SOUVENIR_PACKAGE,
-    LONDON_2018_CACHE_SOUVENIR_PACKAGE,
-    LONDON_2018_TRAIN_SOUVENIR_PACKAGE,
-    LONDON_2018_NUKE_SOUVENIR_PACKAGE,
-    SKILL_GROUPS_CAPSULE,
-    DANGER_ZONE_CASE,
-    OPERATION_SHATTERED_WEB_PREMIUM_PASS,
-    OPERATION_SHATTERED_WEB_CHALLENGE_COIN,
-    SILVER_OPERATION_SHATTERED_WEB_COIN,
-    GOLD_OPERATION_SHATTERED_WEB_COIN,
-    DIAMOND_OPERATION_SHATTERED_WEB_COIN,
-    KATOWICE_2019_VIEWER_PASS,
-    KATOWICE_2019_COIN,
-    KATOWICE_2019_SILVER_COIN,
-    KATOWICE_2019_GOLD_COIN,
-    KATOWICE_2019_DIAMOND_COIN,
-    STICKER_ASTRALIS_KATOWICE_2019,
-    STICKER_AVANGAR_KATOWICE_2019,
-    STICKER_BIG_KATOWICE_2019,
-    STICKER_CLOUD9_KATOWICE_2019,
-    STICKER_COMPLEXITY_GAMING_KATOWICE_2019,
-    STICKER_ENCE_KATOWICE_2019,
-    STICKER_FAZE_CLAN_KATOWICE_2019,
-    STICKER_FNATIC_KATOWICE_2019,
-    STICKER_FURIA_KATOWICE_2019,
-    STICKER_G2_ESPORTS_KATOWICE_2019,
-    STICKER_GRAYHOUND_GAMING_KATOWICE_2019,
-    STICKER_HELLRAISERS_KATOWICE_2019,
-    STICKER_MIBR_KATOWICE_2019,
-    STICKER_NATUS_VINCERE_KATOWICE_2019,
-    STICKER_NINJAS_IN_PYJAMAS_KATOWICE_2019,
-    STICKER_NRG_KATOWICE_2019,
-    STICKER_RENEGADES_KATOWICE_2019,
-    STICKER_TEAM_LIQUID_KATOWICE_2019,
-    STICKER_TEAM_SPIRIT_KATOWICE_2019,
-    STICKER_TYLOO_KATOWICE_2019,
-    STICKER_VEGA_SQUADRON_KATOWICE_2019,
-    STICKER_VICI_GAMING_KATOWICE_2019,
-    STICKER_VITALITY_KATOWICE_2019,
-    STICKER_WINSTRIKE_TEAM_KATOWICE_2019,
-    STICKER_IEM_KATOWICE_2019,
-    KATOWICE_2019_LEGENDS_HOLO_FOIL,
-    KATOWICE_2019_RETURNING_CHALLENGERS_HOLO_FOIL,
-    KATOWICE_2019_MINOR_CHALLENGERS_HOLO_FOIL,
-    KATOWICE_2019_LEGENDS_AUTOGRAPH_CAPSULE,
-    KATOWICE_2019_RETURNING_CHALLENGERS_AUTOGRAPH_CAPSULE,
-    KATOWICE_2019_MINOR_CHALLENGERS_AUTOGRAPH_CAPSULE,
-    KATOWICE_2019_INFERNO_SOUVENIR_PACKAGE,
-    KATOWICE_2019_MIRAGE_SOUVENIR_PACKAGE,
-    KATOWICE_2019_DUST_II_SOUVENIR_PACKAGE,
-    KATOWICE_2019_OVERPASS_SOUVENIR_PACKAGE,
-    KATOWICE_2019_CACHE_SOUVENIR_PACKAGE,
-    KATOWICE_2019_TRAIN_SOUVENIR_PACKAGE,
-    KATOWICE_2019_NUKE_SOUVENIR_PACKAGE,
-    HALO_CAPSULE,
-    PRISMA_CASE,
-    FERAL_PREDATORS_CAPSULE,
-    SHATTERED_WEB_STICKER_COLLECTION,
-    THE_OVERPASS_COLLECTION,
-    THE_COBBLESTONE_COLLECTION,
-    THE_CACHE_COLLECTION,
-    THE_GODS_AND_MONSTERS_COLLECTION,
-    THE_CHOP_SHOP_COLLECTION,
-    THE_RISING_SUN_COLLECTION,
-    BONUS_RANK_XP,
-    THE_CANALS_COLLECTION,
-    PATCH,
-    CSGO_PATCH_PACK,
-    THE_NORSE_COLLECTION,
-    THE_ST_MARC_COLLECTION,
-    BLOODY_DARRYL_THE_STRAPPED_THE_PROFESSIONALS,
-    METAL_SKILL_GROUP_PATCH_COLLECTION,
-    HALF_LIFE_ALYX_PATCH_PACK,
-    WARHAMMER_40000_STICKER_CAPSULE,
-    CSGO_GRAFFITI_3_COLLECTION,
-    TROLLING_GRAFFITI_COLLECTION,
-    BLUEBERRIES_BUCKSHOT_NSWC_SEAL,
-    SHATTERED_WEB_CASE,
-    CSGO_GRAFFITI_2_COLLECTION,
-    BERLIN_2019_VIEWER_PASS,
-    BERLIN_2019_COIN,
-    BERLIN_2019_SILVER_COIN,
-    BERLIN_2019_GOLD_COIN,
-    BERLIN_2019_DIAMOND_COIN,
-    BERLIN_2019_VIEWER_PASS_and_3_SOUVENIR_TOKENS,
-    BERLIN_2019_SOUVENIR_TOKEN,
-    STICKER_ASTRALIS_BERLIN_2019,
-    STICKER_ENCE_BERLIN_2019,
-    STICKER_MIBR_BERLIN_2019,
-    STICKER_NATUS_VINCERE_BERLIN_2019,
-    STICKER_NINJAS_IN_PYJAMAS_BERLIN_2019,
-    STICKER_FAZE_CLAN_BERLIN_2019,
-    STICKER_TEAM_LIQUID_BERLIN_2019,
-    STICKER_RENEGADES_BERLIN_2019,
-    STICKER_COMPLEXITY_GAMING_BERLIN_2019,
-    STICKER_HELLRAISERS_BERLIN_2019,
-    STICKER_AVANGAR_BERLIN_2019,
-    STICKER_G2_ESPORTS_BERLIN_2019,
-    STICKER_VITALITY_BERLIN_2019,
-    STICKER_GRAYHOUND_GAMING_BERLIN_2019,
-    STICKER_MOUSESPORTS_BERLIN_2019,
-    STICKER_FORZE_ESPORTS_BERLIN_2019,
-    STICKER_NRG_BERLIN_2019,
-    STICKER_TYLOO_BERLIN_2019,
-    STICKER_FURIA_BERLIN_2019,
-    STICKER_CR4ZY_BERLIN_2019,
-    STICKER_SYMAN_GAMING_BERLIN_2019,
-    STICKER_NORTH_BERLIN_2019,
-    STICKER_DREAMEATERS_BERLIN_2019,
-    STICKER_INTZ_E_SPORTS_CLUB_BERLIN_2019,
-    STICKER_STARLADDER_BERLIN_2019,
-    BERLIN_2019_LEGENDS_HOLO_FOIL,
-    BERLIN_2019_RETURNING_CHALLENGERS_HOLO_FOIL,
-    BERLIN_2019_MINOR_CHALLENGERS_HOLO_FOIL,
-    BERLIN_2019_LEGENDS_AUTOGRAPH_CAPSULE,
-    BERLIN_2019_RETURNING_CHALLENGERS_AUTOGRAPH_CAPSULE,
-    BERLIN_2019_MINOR_CHALLENGERS_AUTOGRAPH_CAPSULE,
-    BERLIN_2019_INFERNO_SOUVENIR_PACKAGE,
-    BERLIN_2019_MIRAGE_SOUVENIR_PACKAGE,
-    BERLIN_2019_DUST_II_SOUVENIR_PACKAGE,
-    BERLIN_2019_OVERPASS_SOUVENIR_PACKAGE,
-    BERLIN_2019_TRAIN_SOUVENIR_PACKAGE,
-    BERLIN_2019_NUKE_SOUVENIR_PACKAGE,
-    BERLIN_2019_VERTIGO_SOUVENIR_PACKAGE,
-    SHATTERED_WEB_STICKER_COLLECTION1,
-    X_RAY_P250_PACKAGE,
-    CS20_CASE,
-    CS20_STICKER_CAPSULE,
-    _1_STAR_FOR_OPERATION_SHATTERED_WEB,
-    _10_STARS_FOR_OPERATION_SHATTERED_WEB,
-    _100_STARS_FOR_OPERATION_SHATTERED_WEB,
-    _2020_SERVICE_MEDAL,
-    _2020_SERVICE_MEDAL1,
-    _2020_SERVICE_MEDAL2,
-    _2020_SERVICE_MEDAL3,
-    _2020_SERVICE_MEDAL4,
-    _2020_SERVICE_MEDAL5,
-    TWO_TIMES_MCCOY_TACP_CAVALRY,
-    CIVIL_PROTECTION_PIN = 4682,
-    SUSTENANCE_PIN,
-    VORTIGAUNT_PIN,
-    HEADCRAB_GLYPH_PIN,
-    HEALTH_PIN,
-    LAMBDA_PIN,
-    COPPER_LAMBDA_PIN,
-    CMB_PIN,
-    BLACK_MESA_PIN,
-    COMBINE_HELMET_PIN,
-    CITY_17_PIN,
-    HALF_LIFE_ALYX_COLLECTIBLE_PINS_CAPSULE,
-    HALF_LIFE_ALYX_STICKER_CAPSULE,
-    PRISMA_2_CASE,
-    MASTERMINDS_MUSIC_KIT_BOX,
-    STATTRAK_MASTERMINDS_MUSIC_KIT_BOX,
-    FRACTURE_CASE,
-    OPERATION_BROKEN_FANG_PREMIUM_PASS,
-    OPERATION_BROKEN_FANG_CHALLENGE_COIN,
-    SILVER_OPERATION_BROKEN_FANG_COIN,
-    GOLD_OPERATION_BROKEN_FANG_COIN,
-    DIAMOND_OPERATION_BROKEN_FANG_COIN,
-    _1_STAR_FOR_OPERATION_BROKEN_FANG,
-    _10_STARS_FOR_OPERATION_BROKEN_FANG,
-    _100_STARS_FOR_OPERATION_BROKEN_FANG,
-    DISTINGUISHED_AGENTS,
-    EXCEPTIONAL_AGENTS,
-    SUPERIOR_AGENTS,
-    MASTER_AGENT__COMMANDER_MAE,
-    CMDR_MAE_DEAD_COLD_JAMISON_SWAT,
-    _1ST_LIEUTENANT_FARLOW_SWAT,
-    JOHN_VAN_HEALEN_KASK_SWAT,
-    BIO_HAZ_SPECIALIST_SWAT,
-    SERGEANT_BOMBSON_SWAT,
-    CHEM_HAZ_SPECIALIST_SWAT,
-    OPERATION_BROKEN_FANG_CASE,
-    REZAN_THE_REDSHIRT_SABRE,
-    RECOIL_STICKER_COLLECTION,
-    THE_CONTROL_COLLECTION,
-    THE_HAVOC_COLLECTION,
-    THE_ANCIENT_COLLECTION,
-    RECOIL_STICKER_COLLECTION1,
-    MASTER_AGENTS__SIR_BLOODY_DARRYL,
-    BROKEN_FANG_GLOVES,
-    SIR_BLOODY_MIAMI_DARRYL_THE_PROFESSIONALS,
-    SAFECRACKER_VOLTZMANN_THE_PROFESSIONALS,
-    LITTLE_KEV_THE_PROFESSIONALS,
-    BROKEN_FANG_STICKER_COLLECTION,
-    GETAWAY_SALLY_THE_PROFESSIONALS,
-    BROKEN_FANG_STICKER_COLLECTION1,
-    NUMBER_K_THE_PROFESSIONALS,
-    SIR_BLOODY_SILENT_DARRYL_THE_PROFESSIONALS,
-    SIR_BLOODY_SKULLHEAD_DARRYL_THE_PROFESSIONALS,
-    SIR_BLOODY_DARRYL_ROYALE_THE_PROFESSIONALS,
-    SIR_BLOODY_LOUDMOUTH_DARRYL_THE_PROFESSIONALS,
-    _2021_SERVICE_MEDAL,
-    _2021_SERVICE_MEDAL1,
-    _2021_SERVICE_MEDAL2,
-    _2021_SERVICE_MEDAL3,
-    _2021_SERVICE_MEDAL4,
-    _2021_SERVICE_MEDAL5,
-    _2020_RMR_LEGENDS,
-    _2020_RMR_CHALLENGERS,
-    _2020_RMR_CONTENDERS,
-    POORLY_DRAWN_CAPSULE,
-    SNAKEBITE_CASE,
-    COUNTER_STRIKE_SUBSCRIPTION_SERVICE,
-    SOUS_LIEUTENANT_MEDIC_GENDARMERIE_NATIONALE,
-    CHEM_HAZ_CAPITAINE_GENDARMERIE_NATIONALE,
-    CHEF_DESCADRON_ROUCHARD_GENDARMERIE_NATIONALE,
-    ASPIRANT_GENDARMERIE_NATIONALE,
-    OFFICER_JACQUES_BELTRAM_GENDARMERIE_NATIONALE,
-    TACTICIANS_MUSIC_KIT_BOX,
-    STATTRAK_TACTICIANS_MUSIC_KIT_BOX,
-    LIEUTENANT_TREE_HUGGER_FARLOW_SWAT,
-    CMDR_DAVIDA_GOGGLES_FERNANDEZ_SEAL_FROGMAN,
-    OPERATION_RIPTIDE_PREMIUM_PASS,
-    OPERATION_RIPTIDE_CHALLENGE_COIN,
-    SILVER_OPERATION_RIPTIDE_COIN,
-    GOLD_OPERATION_RIPTIDE_COIN,
-    DIAMOND_OPERATION_RIPTIDE_COIN,
-    _1_STAR_FOR_OPERATION_RIPTIDE,
-    _10_STARS_FOR_OPERATION_RIPTIDE,
-    _100_STARS_FOR_OPERATION_RIPTIDE,
-    DISTINGUISHED_AGENTS1,
-    EXCEPTIONAL_AGENTS1,
-    SUPERIOR_AGENTS1,
-    CT_MASTER_AGENTS,
-    T_MASTER_AGENTS,
-    CMDR_FRANK_WET_SOX_BAROUD_SEAL_FROGMAN,
-    LIEUTENANT_REX_KRIKEY_SEAL_FROGMAN,
-    ELITE_TRAPPER_SOLMAN_GUERRILLA_WARFARE,
-    CRASSWATER_THE_FORGOTTEN_GUERRILLA_WARFARE,
-    ARNO_THE_OVERGROWN_GUERRILLA_WARFARE,
-    COL_MANGOS_DABISI_GUERRILLA_WARFARE,
-    VYPA_SISTA_OF_THE_REVOLUTION_GUERRILLA_WARFARE,
-    TRAPPER_AGGRESSOR_GUERRILLA_WARFARE,
-    RIPTIDE_SURF_SHOP_STICKER_COLLECTION,
-    MEDIUM_RARE_CRASSWATER_GUERRILLA_WARFARE,
-    TRAPPER_GUERRILLA_WARFARE,
-    _2021_COMMUNITY_STICKER_CAPSULE,
-    OPERATION_RIPTIDE_STICKER_COLLECTION,
-    THE_BOARDROOM_STICKER_CAPSULE,
-    THE_2021_TRAIN_COLLECTION,
-    THE_2021_TRAIN_COLLECTION1,
-    THE_2021_TRAIN_COLLECTION2,
-    THE_2021_TRAIN_COLLECTION3,
-    BATTLEFIELD_2042_STICKER_CAPSULE,
-    OPERATION_RIPTIDE_CASE,
-    RIPTIDE_SURF_SHOP_STICKER_COLLECTION1,
-    THE_2021_VERTIGO_COLLECTION,
-    THE_2021_DUST_2_COLLECTION,
-    THE_2021_MIRAGE_COLLECTION,
-    OPERATION_RIPTIDE_PATCH_COLLECTION,
-    STOCKHOLM_2021_VIEWER_PASS,
-    STOCKHOLM_2021_COIN,
-    STOCKHOLM_2021_SILVER_COIN,
-    STOCKHOLM_2021_GOLD_COIN,
-    STOCKHOLM_2021_DIAMOND_COIN,
-    STOCKHOLM_2021_VIEWER_PASS_and_3_SOUVENIR_TOKENS,
-    STOCKHOLM_2021_SOUVENIR_TOKEN,
-    STOCKHOLM_2021_LEGENDS_STICKER_CAPSULE,
-    STOCKHOLM_2021_CHALLENGERS_STICKER_CAPSULE,
-    STOCKHOLM_2021_CONTENDERS_STICKER_CAPSULE,
-    STOCKHOLM_2021_LEGENDS_PATCH_PACK,
-    STOCKHOLM_2021_CHALLENGERS_PATCH_PACK,
-    STOCKHOLM_2021_CONTENDERS_PATCH_PACK,
-    STOCKHOLM_2021_INFERNO_SOUVENIR_PACKAGE,
-    STOCKHOLM_2021_MIRAGE_SOUVENIR_PACKAGE,
-    STOCKHOLM_2021_DUST_II_SOUVENIR_PACKAGE,
-    STOCKHOLM_2021_OVERPASS_SOUVENIR_PACKAGE,
-    STOCKHOLM_2021_ANCIENT_SOUVENIR_PACKAGE,
-    STOCKHOLM_2021_NUKE_SOUVENIR_PACKAGE,
-    STOCKHOLM_2021_VERTIGO_SOUVENIR_PACKAGE,
-    STOCKHOLM_2021_CHAMPIONS_AUTOGRAPH_CAPSULE,
-    STOCKHOLM_2021_FINALISTS_AUTOGRAPH_CAPSULE,
-    DREAMS_AND_NIGHTMARES_CASE,
-    _2022_SERVICE_MEDAL,
-    _2022_SERVICE_MEDAL1,
-    _2022_SERVICE_MEDAL2,
-    _2022_SERVICE_MEDAL3,
-    _2022_SERVICE_MEDAL4,
-    _2022_SERVICE_MEDAL5,
-    ANTWERP_2022_VIEWER_PASS,
-    ANTWERP_2022_COIN,
-    ANTWERP_2022_SILVER_COIN,
-    ANTWERP_2022_GOLD_COIN,
-    ANTWERP_2022_DIAMOND_COIN,
-    ANTWERP_2022_VIEWER_PASS_and_3_SOUVENIR_TOKENS,
-    ANTWERP_2022_SOUVENIR_TOKEN,
-    ANTWERP_2022_LEGENDS_STICKER_CAPSULE,
-    ANTWERP_2022_CHALLENGERS_STICKER_CAPSULE,
-    ANTWERP_2022_CONTENDERS_STICKER_CAPSULE,
-    ANTWERP_2022_INFERNO_SOUVENIR_PACKAGE,
-    ANTWERP_2022_MIRAGE_SOUVENIR_PACKAGE,
-    ANTWERP_2022_DUST_II_SOUVENIR_PACKAGE,
-    ANTWERP_2022_OVERPASS_SOUVENIR_PACKAGE,
-    ANTWERP_2022_ANCIENT_SOUVENIR_PACKAGE,
-    ANTWERP_2022_NUKE_SOUVENIR_PACKAGE,
-    ANTWERP_2022_VERTIGO_SOUVENIR_PACKAGE,
-    ANTWERP_2022_LEGENDS_AUTOGRAPH_CAPSULE,
-    ANTWERP_2022_CHALLENGERS_AUTOGRAPH_CAPSULE,
-    ANTWERP_2022_CONTENDERS_AUTOGRAPH_CAPSULE,
-    ANTWERP_2022_CHAMPIONS_AUTOGRAPH_CAPSULE,
-    RECOIL_CASE,
-    _10_YEAR_BIRTHDAY_STICKER_CAPSULE,
-    INITIATORS_MUSIC_KIT_BOX,
-    STATTRAK_INITIATORS_MUSIC_KIT_BOX,
-    RIO_2022_VIEWER_PASS,
-    RIO_2022_COIN,
-    RIO_2022_SILVER_COIN,
-    RIO_2022_GOLD_COIN,
-    RIO_2022_DIAMOND_COIN,
-    RIO_2022_VIEWER_PASS_and_3_SOUVENIR_TOKENS,
-    RIO_2022_SOUVENIR_TOKEN,
-    RIO_2022_LEGENDS_STICKER_CAPSULE,
-    RIO_2022_CHALLENGERS_STICKER_CAPSULE,
-    RIO_2022_CONTENDERS_STICKER_CAPSULE,
-    RIO_2022_INFERNO_SOUVENIR_PACKAGE,
-    RIO_2022_MIRAGE_SOUVENIR_PACKAGE,
-    RIO_2022_DUST_II_SOUVENIR_PACKAGE,
-    RIO_2022_OVERPASS_SOUVENIR_PACKAGE,
-    RIO_2022_ANCIENT_SOUVENIR_PACKAGE,
-    RIO_2022_NUKE_SOUVENIR_PACKAGE,
-    RIO_2022_VERTIGO_SOUVENIR_PACKAGE,
-    RIO_2022_LEGENDS_AUTOGRAPH_CAPSULE,
-    RIO_2022_CHALLENGERS_AUTOGRAPH_CAPSULE,
-    RIO_2022_CONTENDERS_AUTOGRAPH_CAPSULE,
-    RIO_2022_CHAMPIONS_AUTOGRAPH_CAPSULE,
-    RIO_2022_STORAGE_UNIT_WITH_PAPER_STICKERS,
-    RIO_2022_STORAGE_UNIT_WITH_GLITTER_STICKERS,
-    _2023_SERVICE_MEDAL,
-    _2023_SERVICE_MEDAL1,
-    _2023_SERVICE_MEDAL2,
-    _2023_SERVICE_MEDAL3,
-    _2023_SERVICE_MEDAL4,
-    _2023_SERVICE_MEDAL5,
-    ESPIONAGE_STICKER_CAPSULE,
-    REVOLUTION_CASE,
-    ANUBIS_COLLECTION_PACKAGE = 4882,
-    PARIS_2023_VIEWER_PASS,
-    PARIS_2023_COIN,
-    PARIS_2023_SILVER_COIN,
-    PARIS_2023_GOLD_COIN,
-    PARIS_2023_DIAMOND_COIN,
-    PARIS_2023_VIEWER_PASS_and_3_SOUVENIR_TOKENS,
-    PARIS_2023_SOUVENIR_TOKEN,
-    PARIS_2023_LEGENDS_STICKER_CAPSULE,
-    PARIS_2023_CHALLENGERS_STICKER_CAPSULE,
-    PARIS_2023_CONTENDERS_STICKER_CAPSULE,
-    PARIS_2023_INFERNO_SOUVENIR_PACKAGE,
-    PARIS_2023_MIRAGE_SOUVENIR_PACKAGE,
-    PARIS_2023_ANUBIS_SOUVENIR_PACKAGE,
-    PARIS_2023_OVERPASS_SOUVENIR_PACKAGE,
-    PARIS_2023_ANCIENT_SOUVENIR_PACKAGE,
-    PARIS_2023_NUKE_SOUVENIR_PACKAGE,
-    PARIS_2023_VERTIGO_SOUVENIR_PACKAGE,
-    PARIS_2023_LEGENDS_AUTOGRAPH_CAPSULE,
-    PARIS_2023_CHALLENGERS_AUTOGRAPH_CAPSULE,
-    PARIS_2023_CONTENDERS_AUTOGRAPH_CAPSULE,
-    PARIS_2023_CHAMPIONS_AUTOGRAPH_CAPSULE,
-    BLOODHOUND_GLOVES = 5027,
-    DEFAULT_T_GLOVES,
-    DEFAULT_CT_GLOVES,
-    SPORT_GLOVES,
-    DRIVER_GLOVES,
-    HAND_WRAPS,
-    MOTO_GLOVES,
-    SPECIALIST_GLOVES,
-    HYDRA_GLOVES,
-    DEFAULT_T_AGENT,
-    DEFAULT_CT_AGENT,
-    CSGO_CUSTOMPLAYER_TM_ANARCHIST,
-    CSGO_CUSTOMPLAYER_TM_ANARCHIST_VARIANTA,
-    CSGO_CUSTOMPLAYER_TM_ANARCHIST_VARIANTB,
-    CSGO_CUSTOMPLAYER_TM_ANARCHIST_VARIANTC,
-    CSGO_CUSTOMPLAYER_TM_ANARCHIST_VARIANTD,
-    CSGO_CUSTOMPLAYER_TM_PIRATE,
-    CSGO_CUSTOMPLAYER_TM_PIRATE_VARIANTA,
-    CSGO_CUSTOMPLAYER_TM_PIRATE_VARIANTB,
-    CSGO_CUSTOMPLAYER_TM_PIRATE_VARIANTC,
-    CSGO_CUSTOMPLAYER_TM_PIRATE_VARIANTD,
-    CSGO_CUSTOMPLAYER_TM_PROFESSIONAL,
-    CSGO_CUSTOMPLAYER_TM_PROFESSIONAL_VAR1,
-    CSGO_CUSTOMPLAYER_TM_PROFESSIONAL_VAR2,
-    CSGO_CUSTOMPLAYER_TM_PROFESSIONAL_VAR3,
-    CSGO_CUSTOMPLAYER_TM_PROFESSIONAL_VAR4,
-    CSGO_CUSTOMPLAYER_TM_SEPARATIST,
-    CSGO_CUSTOMPLAYER_TM_SEPARATIST_VARIANTA,
-    CSGO_CUSTOMPLAYER_TM_SEPARATIST_VARIANTB,
-    CSGO_CUSTOMPLAYER_TM_SEPARATIST_VARIANTC,
-    CSGO_CUSTOMPLAYER_TM_SEPARATIST_VARIANTD,
-    CSGO_CUSTOMPLAYER_CTM_GIGN,
-    CSGO_CUSTOMPLAYER_CTM_GIGN_VARIANTA,
-    CSGO_CUSTOMPLAYER_CTM_GIGN_VARIANTB,
-    CSGO_CUSTOMPLAYER_CTM_GIGN_VARIANTC,
-    CSGO_CUSTOMPLAYER_CTM_GIGN_VARIANTD,
-    CSGO_CUSTOMPLAYER_CTM_GSG9,
-    CSGO_CUSTOMPLAYER_CTM_GSG9_VARIANTA,
-    CSGO_CUSTOMPLAYER_CTM_GSG9_VARIANTB,
-    CSGO_CUSTOMPLAYER_CTM_GSG9_VARIANTC,
-    CSGO_CUSTOMPLAYER_CTM_GSG9_VARIANTD,
-    CSGO_CUSTOMPLAYER_CTM_IDF,
-    CSGO_CUSTOMPLAYER_CTM_IDF_VARIANTB,
-    CSGO_CUSTOMPLAYER_CTM_IDF_VARIANTC,
-    CSGO_CUSTOMPLAYER_CTM_IDF_VARIANTD,
-    CSGO_CUSTOMPLAYER_CTM_IDF_VARIANTE,
-    CSGO_CUSTOMPLAYER_CTM_IDF_VARIANTF,
-    CSGO_CUSTOMPLAYER_CTM_SWAT,
-    CSGO_CUSTOMPLAYER_CTM_SWAT_VARIANTA,
-    CSGO_CUSTOMPLAYER_CTM_SWAT_VARIANTB,
-    CSGO_CUSTOMPLAYER_CTM_SWAT_VARIANTC,
-    CSGO_CUSTOMPLAYER_CTM_SWAT_VARIANTD,
-    CSGO_CUSTOMPLAYER_CTM_SAS_VARIANTA,
-    CSGO_CUSTOMPLAYER_CTM_SAS_VARIANTB,
-    CSGO_CUSTOMPLAYER_CTM_SAS_VARIANTC,
-    CSGO_CUSTOMPLAYER_CTM_SAS_VARIANTD,
-    CSGO_CUSTOMPLAYER_CTM_ST6,
-    CSGO_CUSTOMPLAYER_CTM_ST6_VARIANTA,
-    CSGO_CUSTOMPLAYER_CTM_ST6_VARIANTB,
-    CSGO_CUSTOMPLAYER_CTM_ST6_VARIANTC,
-    CSGO_CUSTOMPLAYER_CTM_ST6_VARIANTD,
-    CSGO_CUSTOMPLAYER_TM_BALKAN_VARIANTE,
-    CSGO_CUSTOMPLAYER_TM_BALKAN_VARIANTA,
-    CSGO_CUSTOMPLAYER_TM_BALKAN_VARIANTB,
-    CSGO_CUSTOMPLAYER_TM_BALKAN_VARIANTC,
-    CSGO_CUSTOMPLAYER_TM_BALKAN_VARIANTD,
-    CSGO_CUSTOMPLAYER_TM_JUMPSUIT_VARIANTA,
-    CSGO_CUSTOMPLAYER_TM_JUMPSUIT_VARIANTB,
-    CSGO_CUSTOMPLAYER_TM_JUMPSUIT_VARIANTC,
-    CSGO_CUSTOMPLAYER_TM_PHOENIX_HEAVY,
-    CSGO_CUSTOMPLAYER_CTM_HEAVY,
-    CSGO_CUSTOMPLAYER_TM_LEET_VARIANTA = 5100,
-    CSGO_CUSTOMPLAYER_TM_LEET_VARIANTB,
-    CSGO_CUSTOMPLAYER_TM_LEET_VARIANTC,
-    CSGO_CUSTOMPLAYER_TM_LEET_VARIANTD,
-    CSGO_CUSTOMPLAYER_TM_LEET_VARIANTE,
-    GROUND_REBEL__ELITE_CREW,
-    OSIRIS_ELITE_CREW,
-    PROF_SHAHMAT_ELITE_CREW,
-    THE_ELITE_MR_MUHLIK_ELITE_CREW,
-    JUNGLE_REBEL_ELITE_CREW,
-    CSGO_CUSTOMPLAYER_TM_PHOENIX = 5200,
-    CSGO_CUSTOMPLAYER_TM_PHOENIX_VARIANTA,
-    CSGO_CUSTOMPLAYER_TM_PHOENIX_VARIANTB,
-    CSGO_CUSTOMPLAYER_TM_PHOENIX_VARIANTC,
-    CSGO_CUSTOMPLAYER_TM_PHOENIX_VARIANTD,
-    SOLDIER_PHOENIX,
-    ENFORCER_PHOENIX,
-    SLINGSHOT_PHOENIX,
-    STREET_SOLDIER_PHOENIX,
-    CSGO_CUSTOMPLAYER_CTM_FBI = 5300,
-    CSGO_CUSTOMPLAYER_CTM_FBI_VARIANTA,
-    CSGO_CUSTOMPLAYER_CTM_FBI_VARIANTC,
-    CSGO_CUSTOMPLAYER_CTM_FBI_VARIANTD,
-    CSGO_CUSTOMPLAYER_CTM_FBI_VARIANTE,
-    OPERATOR_FBI_SWAT,
-    MARKUS_DELROW_FBI_HRT,
-    MICHAEL_SYFERS__FBI_SNIPER,
-    SPECIAL_AGENT_AVA_FBI,
-    _3RD_COMMANDO_COMPANY_KSK = 5400,
-    SEAL_TEAM_6_SOLDIER_NSWC_SEAL,
-    BUCKSHOT_NSWC_SEAL,
-    TWO_TIMES_MCCOY_USAF_TACP,
-    LT_COMMANDER_RICKSAW_NSWC_SEAL,
-    PRIMEIRO_TENENTE_BRAZILIAN_1ST_BATTALION,
-    DRAGOMIR_SABRE = 5500,
-    MAXIMUS_SABRE,
-    REZAN_THE_READY_SABRE,
-    BLACKWOLF_SABRE,
-    THE_DOCTOR_ROMANOV_SABRE,
-    DRAGOMIR_SABRE_FOOTSOLDIER,
-    CSGO_CUSTOMPLAYER_CTM_SAS = 5600,
-    B_SQUADRON_OFFICER_SAS,
-    D_SQUADRON_OFFICER_NZSAS,
-    DUST_II_PIN = 6001,
-    GUARDIAN_ELITE_PIN,
-    MIRAGE_PIN,
-    INFERNO_PIN,
-    ITALY_PIN,
-    VICTORY_PIN,
-    MILITIA_PIN,
-    NUKE_PIN,
-    TRAIN_PIN,
-    GUARDIAN_PIN,
-    TACTICS_PIN,
-    GUARDIAN_2_PIN,
-    BRAVO_PIN,
-    BAGGAGE_PIN,
-    PHOENIX_PIN,
-    OFFICE_PIN,
-    COBBLESTONE_PIN,
-    OVERPASS_PIN,
-    BLOODHOUND_PIN,
-    CACHE_PIN,
-    VALERIA_PHOENIX_PIN,
-    CHROMA_PIN,
-    GUARDIAN_3_PIN,
-    CANALS_PIN,
-    WELCOME_TO_THE_CLUTCH_PIN,
-    DEATH_SENTENCE_PIN,
-    INFERNO_2_PIN,
-    WILDFIRE_PIN,
-    EASY_PEASY_PIN,
-    ACES_HIGH_PIN,
-    HYDRA_PIN,
-    HOWL_PIN,
-    BRIGADIER_GENERAL_PIN,
-    ALYX_PIN,
-    NO_LOCALIZED_NAME1 = 6046,
-    NO_LOCALIZED_NAME2,
-    NO_LOCALIZED_NAME3,
-    NO_LOCALIZED_NAME4,
-    NO_LOCALIZED_NAME5,
-    NO_LOCALIZED_NAME6,
-    NO_LOCALIZED_NAME7,
-    NO_LOCALIZED_NAME8,
-    NO_LOCALIZED_NAME9,
-    NO_LOCALIZED_NAME10,
-    NO_LOCALIZED_NAME11,
-    NO_LOCALIZED_NAME12,
-    NO_LOCALIZED_NAME13,
-    NO_LOCALIZED_NAME14,
-    NO_LOCALIZED_NAME15,
-    NO_LOCALIZED_NAME16,
-    NO_LOCALIZED_NAME17,
-    NO_LOCALIZED_NAME18,
-    NO_LOCALIZED_NAME19,
-    NO_LOCALIZED_NAME20,
-    NO_LOCALIZED_NAME21,
-    NO_LOCALIZED_NAME22,
-    NO_LOCALIZED_NAME23,
-    NO_LOCALIZED_NAME24,
-    NO_LOCALIZED_NAME25,
-    NO_LOCALIZED_NAME26,
-    NO_LOCALIZED_NAME27,
-    NO_LOCALIZED_NAME28,
-    NO_LOCALIZED_NAME29,
-    NO_LOCALIZED_NAME30,
-    NO_LOCALIZED_NAME31,
-    NO_LOCALIZED_NAME32,
-    NO_LOCALIZED_NAME33,
-    NO_LOCALIZED_NAME34,
-    NO_LOCALIZED_NAME35,
-    NO_LOCALIZED_NAME36,
-    NO_LOCALIZED_NAME37,
-    NO_LOCALIZED_NAME38,
-    NO_LOCALIZED_NAME39,
-    NO_LOCALIZED_NAME40,
-    NO_LOCALIZED_NAME41,
-    NO_LOCALIZED_NAME42,
-    NO_LOCALIZED_NAME43,
-    DUST_II_PIN1 = 6101,
-    GUARDIAN_ELITE_PIN1,
-    MIRAGE_PIN1,
-    INFERNO_PIN1,
-    ITALY_PIN1,
-    VICTORY_PIN1,
-    MILITIA_PIN1,
-    NUKE_PIN1,
-    TRAIN_PIN1,
-    GUARDIAN_PIN1,
-    TACTICS_PIN1,
-    GUARDIAN_2_PIN1,
-    BRAVO_PIN1,
-    BAGGAGE_PIN1,
-    PHOENIX_PIN1,
-    OFFICE_PIN1,
-    COBBLESTONE_PIN1,
-    OVERPASS_PIN1,
-    BLOODHOUND_PIN1,
-    CACHE_PIN1,
-    VALERIA_PHOENIX_PIN1,
-    CHROMA_PIN1,
-    GUARDIAN_3_PIN1,
-    CANALS_PIN1,
-    WELCOME_TO_THE_CLUTCH_PIN1,
-    DEATH_SENTENCE_PIN1,
-    INFERNO_2_PIN1,
-    WILDFIRE_PIN1,
-    EASY_PEASY_PIN1,
-    ACES_HIGH_PIN1,
-    HYDRA_PIN1,
-    HOWL_PIN1,
-    BRIGADIER_GENERAL_PIN1,
-    ALYX_PIN1,
-    DISTINGUISHED_AGENTS2 = 6404,
-    EXCEPTIONAL_AGENTS2,
-    SUPERIOR_AGENTS2,
-    MASTER_AGENTS,
-    OPERATION_WILDFIRE_CASE_KEY = 7000,
-    STICKER_BOSSY_BURGER = 20000,
-    STICKER_CAT_CALL,
-    STICKER_CHICKEN_STRIKE,
-    STICKER_CT_IN_BANANA,
-    STICKER_DONT_WORRY_IM_PRO,
-    STICKER_FIGHT_LIKE_A_GIRL,
-    STICKER_FLASHBANG,
-    STICKER_KAWAII_KILLER_CT,
-    STICKER_NELU_THE_BEAR,
-    STICKER_ONE_SHOT_ONE_KILL,
-    STICKER_SHOOTING_STAR_RETURN,
-    STICKER_WAR_PENGUIN = 20012,
-    STICKER_WINDY_WALKING_CLUB,
-    STICKER_BLITZKRIEG,
-    STICKER_PIGEON_MASTER,
-    STICKER_TERRORIZED,
-    STICKER_TILL_DEATH_DO_US_PART,
-    STICKER_STAY_FROSTY,
-    STICKER_T_ON_CAT,
-    MUSIC_KIT_DANIEL_SADOWSKI_CRIMSON_ASSAULT,
-    MUSIC_KIT_NOISIA_SHARPENED,
-    MUSIC_KIT_ROBERT_ALLAIRE_INSURGENCY,
-    MUSIC_KIT_SEAN_MURRAY_AD8,
-    MUSIC_KIT_FEED_ME_HIGH_NOON,
-    MUSIC_KIT_DREN_DEATHS_HEAD_DEMOLITION,
-    MUSIC_KIT_AUSTIN_WINTORY_DESERT_FIRE,
-    MUSIC_KIT_SASHA_LNOE,
-    MUSIC_KIT_SKOG_METAL,
-    STICKER_DOOMED,
-    STICKER_QUEEN_OF_PAIN,
-    STICKER_TRICK_OR_THREAT,
-    STICKER_TRICK_OR_TREAT,
-    STICKER_WITCH,
-    STICKER_ZOMBIE_LOVER,
-    STICKER_BLOOD_BOILER,
-    STICKER_DINKED,
-    STICKER_DRUG_WAR_VETERAN,
-    STICKER_HO_HO_HO,
-    STICKER_MASSIVE_PEAR,
-    STICKER_MY_LITTLE_FRIEND,
-    STICKER_PANDAMONIUM,
-    STICKER_PIECE_OF_CAKE,
-    STICKER_SAS_CHICKEN,
-    STICKER_THUG_LIFE,
-    STICKER_T_REKT,
-    STICKER_WAROWL,
-    STICKER_WORK_FOR_AMMO,
-    STICKER_PHOENIX_FOIL,
-    STICKER_BOMB_SQUAD_FOIL,
-    MUSIC_KIT_MIDNIGHT_RIDERS_ALL_I_WANT_FOR_CHRISTMAS,
-    MUSIC_KIT_DANIEL_SADOWSKI_TOTAL_DOMINATION,
-    MUSIC_KIT_VARIOUS_ARTISTS_HOTLINE_MIAMI,
-    MUSIC_KIT_MATT_LANGE_ISORHYTHM,
-    MUSIC_KIT_MATEO_MESSINA_FOR_NO_MANKIND,
-    MUSIC_KIT_DAMJAN_MRAVUNAC_THE_TALOS_PRINCIPLE,
-    STICKER_FLICKSHOT,
-    STICKER_HEADSHOT_GUARANTEE,
-    STICKER_ECO_RUSH,
-    STICKER_JUST_TROLLING,
-    STICKER_FIRESTARTER_HOLO = 20061,
-    STICKER_LUCKY_CAT_FOIL,
-    STICKER_ROBO,
-    STICKER_WITCHCRAFT,
-    STICKER_WANNA_FIGHT,
-    STICKER_HOSTAGE_RESCUE,
-    STICKER_HAMSTER_HAWK,
-    STICKER_HEADLESS_CHICKEN,
-    MUSIC_KIT_PROXY_BATTLEPACK,
-    MUSIC_KIT_KITHEORY_MOLOTOV,
-    MUSIC_KIT_TROELS_FOLMANN_UBER_BLASTO_PHONE,
-    MUSIC_KIT_KELLY_BAILEY_HAZARDOUS_ENVIRONMENTS,
-    MUSIC_KIT_SKOG_II_HEADSHOT,
-    ENFU_STICKER_CAPSULE1,
-    STICKER_AWP_COUNTRY,
-    STICKER_CHI_BOMB,
-    STICKER_DORU_THE_FOX,
-    STICKER_KNIFE_CLUB,
-    STICKER_CS_ON_THE_MIND,
-    STICKER_NINJA_DEFUSE,
-    STICKER_PROS_DONT_FAKE,
-    STICKER_KAWAII_KILLER_TERRORIST,
-    STICKER_BAAA_CKSTABBER,
-    STICKER_DELICIOUS_TEARS,
-    MUSIC_KIT_DANIEL_SADOWSKI_THE_8_BIT_KIT,
-    MUSIC_KIT_AWOLNATION_I_AM,
-    MUSIC_KIT_MORD_FUSTANG_DIAMONDS,
-    MUSIC_KIT_MICHAEL_BROSS_INVASION,
-    MUSIC_KIT_IAN_HULTQUIST_LIONS_MOUTH,
-    MUSIC_KIT_NEW_BEAT_FUND_SPONGE_FINGERZ,
-    MUSIC_KIT_BEARTOOTH_DISGUSTING,
-    MUSIC_KIT_LENNIE_MOORE_JAVA_HAVANA_FUNKALOO,
-    MUSIC_KIT_DARUDE_MOMENTS_CSGO,
-    STATTRAK_MUSIC_KIT_PROXY_BATTLEPACK,
-    STATTRAK_MUSIC_KIT_KITHEORY_MOLOTOV,
-    STATTRAK_MUSIC_KIT_TROELS_FOLMANN_UBER_BLASTO_PHONE,
-    STATTRAK_MUSIC_KIT_KELLY_BAILEY_HAZARDOUS_ENVIRONMENTS,
-    STATTRAK_MUSIC_KIT_SKOG_II_HEADSHOT,
-    STATTRAK_MUSIC_KIT_DANIEL_SADOWSKI_THE_8_BIT_KIT,
-    STATTRAK_MUSIC_KIT_AWOLNATION_I_AM,
-    STATTRAK_MUSIC_KIT_MORD_FUSTANG_DIAMONDS,
-    STATTRAK_MUSIC_KIT_MICHAEL_BROSS_INVASION,
-    STATTRAK_MUSIC_KIT_IAN_HULTQUIST_LIONS_MOUTH,
-    STATTRAK_MUSIC_KIT_NEW_BEAT_FUND_SPONGE_FINGERZ,
-    STATTRAK_MUSIC_KIT_BEARTOOTH_DISGUSTING,
-    STATTRAK_MUSIC_KIT_LENNIE_MOORE_JAVA_HAVANA_FUNKALOO,
-    STATTRAK_MUSIC_KIT_DARUDE_MOMENTS_CSGO,
-    STATTRAK_MUSIC_KIT_DANIEL_SADOWSKI_CRIMSON_ASSAULT,
-    STATTRAK_MUSIC_KIT_NOISIA_SHARPENED,
-    STATTRAK_MUSIC_KIT_ROBERT_ALLAIRE_INSURGENCY,
-    STATTRAK_MUSIC_KIT_SEAN_MURRAY_AD8,
-    STATTRAK_MUSIC_KIT_FEED_ME_HIGH_NOON,
-    STATTRAK_MUSIC_KIT_DREN_DEATHS_HEAD_DEMOLITION,
-    STATTRAK_MUSIC_KIT_AUSTIN_WINTORY_DESERT_FIRE,
-    STATTRAK_MUSIC_KIT_SASHA_LNOE,
-    STATTRAK_MUSIC_KIT_SKOG_METAL,
-    STATTRAK_MUSIC_KIT_MIDNIGHT_RIDERS_ALL_I_WANT_FOR_CHRISTMAS,
-    STATTRAK_MUSIC_KIT_DANIEL_SADOWSKI_TOTAL_DOMINATION,
-    STATTRAK_MUSIC_KIT_VARIOUS_ARTISTS_HOTLINE_MIAMI,
-    STATTRAK_MUSIC_KIT_MATT_LANGE_ISORHYTHM,
-    STATTRAK_MUSIC_KIT_MATEO_MESSINA_FOR_NO_MANKIND,
-    STATTRAK_MUSIC_KIT_DAMJAN_MRAVUNAC_THE_TALOS_PRINCIPLE,
-    PINUPS_STICKER_CAPSULE,
-    SLID3_STICKER_CAPSULE,
-    TEAM_ROLES_STICKER_CAPSULE,
-    COLLECTIBLE_PINS_CAPSULE_SERIES_11,
-    SUGARFACE_STICKER_CAPSULE,
-    BESTIARY_STICKER_CAPSULE,
-    CSGO_GRAFFITI_BOX1,
-    COMMUNITY_GRAFFITI_BOX_11,
-    COLLECTIBLE_PINS_CAPSULE_SERIES_21,
-    STATTRAK_RADICALS_MUSIC_KIT_BOX = 20133,
-    PERFECT_WORLD_STICKER_CAPSULE_11,
-    PERFECT_WORLD_STICKER_CAPSULE_21,
-    PERFECT_WORLD_GRAFFITI_BOX1,
-    COMMUNITY_CAPSULE_20181,
-    COLLECTIBLE_PINS_CAPSULE_SERIES_31,
-    SKILL_GROUPS_CAPSULE1,
-    MUSIC_KIT_THE_VERKKARS_EZ4ENCE,
-    STATTRAK_MUSIC_KIT_THE_VERKKARS_EZ4ENCE,
-    FERAL_PREDATORS_STICKER_CAPSULE,
-    CHICKEN_CAPSULE1,
-    X_RAY_P250_PACKAGE1,
-    CS20_STICKER_CAPSULE1,
-    HALO_STICKER_CAPSULE,
-    MUSIC_KIT_SCARLXRD_KING_SCAR,
-    STATTRAK_MUSIC_KIT_SCARLXRD_KING_SCAR,
-    CSGO_PATCH_PACK1,
-    HALF_LIFE_ALYX_COLLECTIBLE_PINS_CAPSULE1 = 20152,
-    HALF_LIFE_ALYX_STICKER_CAPSULE1,
-    HALF_LIFE_ALYX_PATCH_PACK1,
-    MASTERMINDS_MUSIC_KIT_BOX1 = 20169,
-    STATTRAK_MASTERMINDS_MUSIC_KIT_BOX1,
-    WARHAMMER_40000_STICKER_CAPSULE1,
-    MUSIC_KIT_AMON_TOBIN_ALL_FOR_DUST,
-    STATTRAK_MUSIC_KIT_AMON_TOBIN_ALL_FOR_DUST,
-    POORLY_DRAWN_STICKER_CAPSULE,
-    MUSIC_KIT_NECK_DEEP_THE_LOWLIFE_PACK,
-    STATTRAK_MUSIC_KIT_NECK_DEEP_THE_LOWLIFE_PACK,
-    MUSIC_KIT_SCARLXRD_CHAINSAWLXADXUT,
-    STATTRAK_MUSIC_KIT_SCARLXRD_CHAINSAWLXADXUT,
-    TACTICIANS_MUSIC_KIT_BOX1,
-    STATTRAK_TACTICIANS_MUSIC_KIT_BOX1,
-    _2021_COMMUNITY_STICKER_CAPSULE1,
-    BATTLEFIELD_2042_STICKER_CAPSULE1,
-    MUSIC_KIT_BBNOS_U_MAD,
-    STATTRAK_MUSIC_KIT_BBNOS_U_MAD,
-    MUSIC_KIT_THE_VERKKARS_AND_N0THING_FLASHBANG_DANCE,
-    STATTRAK_MUSIC_KIT_THE_VERKKARS_AND_N0THING_FLASHBANG_DANCE,
-    THE_BOARDROOM_STICKER_CAPSULE1,
-    _10_YEAR_BIRTHDAY_STICKER_CAPSULE1,
-    INITIATORS_MUSIC_KIT_BOX1,
-    STATTRAK_INITIATORS_MUSIC_KIT_BOX1,
-    MUSIC_KIT_PERFECT_WORLD_HUA_LIAN_PAINTED_FACE,
-    STATTRAK_MUSIC_KIT_PERFECT_WORLD_HUA_LIAN_PAINTED_FACE,
-    ESPIONAGE_STICKER_CAPSULE1,
-    MUSIC_KIT_DENZEL_CURRY_ULTIMATE,
-    STATTRAK_MUSIC_KIT_DENZEL_CURRY_ULTIMATE,
-    ANUBIS_COLLECTION_PACKAGE1 = 20197,
-    DREAMS_AND_NIGHTMARES_CONTEST_WINNER = 30050,
-    ANUBIS_COLLECTION_WORKSHOP_CONTRIBUTOR = 30092,
-};
